@@ -3,18 +3,19 @@ package coolq
 import (
 	"crypto/md5"
 	"encoding/hex"
+	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"math"
 	"os"
 	"path"
 	"path/filepath"
 	"runtime"
-	"runtime/debug"
 	"strconv"
 	"strings"
+	"sync"
 	"time"
-	"unicode/utf8"
+
+	"github.com/segmentio/asm/base64"
 
 	"github.com/Mrs4s/MiraiGo/binary"
 	"github.com/Mrs4s/MiraiGo/client"
@@ -23,36 +24,313 @@ import (
 	log "github.com/sirupsen/logrus"
 	"github.com/tidwall/gjson"
 
+	"github.com/Mrs4s/go-cqhttp/db"
 	"github.com/Mrs4s/go-cqhttp/global"
+	"github.com/Mrs4s/go-cqhttp/internal/base"
+	"github.com/Mrs4s/go-cqhttp/internal/cache"
+	"github.com/Mrs4s/go-cqhttp/internal/param"
+	"github.com/Mrs4s/go-cqhttp/modules/filter"
 )
 
-// Version go-cqhttp的版本信息，在编译时使用ldflags进行覆盖
-var Version = "unknown"
+type guildMemberPageToken struct {
+	guildID        uint64
+	nextIndex      uint32
+	nextRoleID     uint64
+	nextQueryParam string
+}
 
-func init() {
-	if Version != "unknown" {
-		return
-	}
-	info, ok := debug.ReadBuildInfo()
-	if ok {
-		Version = info.Main.Version
-	}
+var defaultPageToken = &guildMemberPageToken{
+	guildID:    0,
+	nextIndex:  0,
+	nextRoleID: 2,
 }
 
 // CQGetLoginInfo 获取登录号信息
 //
 // https://git.io/Jtz1I
-func (bot *CQBot) CQGetLoginInfo() MSG {
-	return OK(MSG{"user_id": bot.Client.Uin, "nickname": bot.Client.Nickname})
+// @route(get_login_info)
+func (bot *CQBot) CQGetLoginInfo() global.MSG {
+	return OK(global.MSG{"user_id": bot.Client.Uin, "nickname": bot.Client.Nickname})
+}
+
+// CQGetQiDianAccountInfo 获取企点账号信息
+// @route(qidian_get_account_info)
+func (bot *CQBot) CQGetQiDianAccountInfo() global.MSG {
+	if bot.Client.QiDian == nil {
+		return Failed(100, "QIDIAN_PROTOCOL_REQUEST", "请使用企点协议")
+	}
+	return OK(global.MSG{
+		"master_id":   bot.Client.QiDian.MasterUin,
+		"ext_name":    bot.Client.QiDian.ExtName,
+		"create_time": bot.Client.QiDian.CreateTime,
+	})
+}
+
+// CQGetGuildServiceProfile 获取频道系统个人资料
+// @route(get_guild_service_profile)
+func (bot *CQBot) CQGetGuildServiceProfile() global.MSG {
+	return OK(global.MSG{
+		"nickname":   bot.Client.GuildService.Nickname,
+		"tiny_id":    fU64(bot.Client.GuildService.TinyId),
+		"avatar_url": bot.Client.GuildService.AvatarUrl,
+	})
+}
+
+// CQGetGuildList 获取已加入的频道列表
+// @route(get_guild_list)
+func (bot *CQBot) CQGetGuildList() global.MSG {
+	fs := make([]global.MSG, 0, len(bot.Client.GuildService.Guilds))
+	for _, info := range bot.Client.GuildService.Guilds {
+		/* 做成单独的 api 可能会好些?
+		channels := make([]global.MSG, 0, len(info.Channels))
+		for _, channel := range info.Channels {
+			channels = append(channels, global.MSG{
+				"channel_id":   channel.ChannelId,
+				"channel_name": channel.ChannelName,
+				"channel_type": channel.ChannelType,
+			})
+		}
+		*/
+		fs = append(fs, global.MSG{
+			"guild_id":         fU64(info.GuildId),
+			"guild_name":       info.GuildName,
+			"guild_display_id": fU64(info.GuildCode),
+			// "channels":         channels,
+		})
+	}
+	return OK(fs)
+}
+
+// CQGetGuildMetaByGuest 通过访客权限获取频道元数据
+// @route(get_guild_meta_by_guest)
+func (bot *CQBot) CQGetGuildMetaByGuest(guildID uint64) global.MSG {
+	meta, err := bot.Client.GuildService.FetchGuestGuild(guildID)
+	if err != nil {
+		log.Errorf("获取频道元数据时出现错误: %v", err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	return OK(global.MSG{
+		"guild_id":         fU64(meta.GuildId),
+		"guild_name":       meta.GuildName,
+		"guild_profile":    meta.GuildProfile,
+		"create_time":      meta.CreateTime,
+		"max_member_count": meta.MaxMemberCount,
+		"max_robot_count":  meta.MaxRobotCount,
+		"max_admin_count":  meta.MaxAdminCount,
+		"member_count":     meta.MemberCount,
+		"owner_id":         fU64(meta.OwnerId),
+	})
+}
+
+// CQGetGuildChannelList 获取频道列表
+// @route(get_guild_channel_list)
+func (bot *CQBot) CQGetGuildChannelList(guildID uint64, noCache bool) global.MSG {
+	guild := bot.Client.GuildService.FindGuild(guildID)
+	if guild == nil {
+		return Failed(100, "GUILD_NOT_FOUND")
+	}
+	if noCache {
+		channels, err := bot.Client.GuildService.FetchChannelList(guildID)
+		if err != nil {
+			log.Warnf("获取频道 %v 子频道列表时出现错误: %v", guildID, err)
+			return Failed(100, "API_ERROR", err.Error())
+		}
+		guild.Channels = channels
+	}
+	channels := make([]global.MSG, 0, len(guild.Channels))
+	for _, c := range guild.Channels {
+		channels = append(channels, convertChannelInfo(c))
+	}
+	return OK(channels)
+}
+
+// CQGetGuildMembers 获取频道成员列表
+// @route(get_guild_member_list)
+func (bot *CQBot) CQGetGuildMembers(guildID uint64, nextToken string) global.MSG {
+	guild := bot.Client.GuildService.FindGuild(guildID)
+	if guild == nil {
+		return Failed(100, "GUILD_NOT_FOUND")
+	}
+	token := defaultPageToken
+	if nextToken != "" {
+		i, exists := bot.nextTokenCache.Get(nextToken)
+		if !exists {
+			return Failed(100, "NEXT_TOKEN_NOT_EXISTS")
+		}
+		token = i.(*guildMemberPageToken)
+		if token.guildID != guildID {
+			return Failed(100, "GUILD_NOT_MATCH")
+		}
+	}
+	ret, err := bot.Client.GuildService.FetchGuildMemberListWithRole(guildID, 0, token.nextIndex, token.nextRoleID, token.nextQueryParam)
+	if err != nil {
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	res := global.MSG{
+		"members":    convertGuildMemberInfo(ret.Members),
+		"finished":   ret.Finished,
+		"next_token": nil,
+	}
+	if !ret.Finished {
+		next := &guildMemberPageToken{
+			guildID:        guildID,
+			nextIndex:      ret.NextIndex,
+			nextRoleID:     ret.NextRoleId,
+			nextQueryParam: ret.NextQueryParam,
+		}
+		id := base64.StdEncoding.EncodeToString(binary.NewWriterF(func(w *binary.Writer) {
+			w.WriteUInt64(uint64(time.Now().UnixNano()))
+			w.WriteString(utils.RandomString(5))
+		}))
+		bot.nextTokenCache.Add(id, next, time.Minute*10)
+		res["next_token"] = id
+	}
+	return OK(res)
+}
+
+// CQGetGuildMemberProfile 获取频道成员资料
+// @route(get_guild_member_profile)
+func (bot *CQBot) CQGetGuildMemberProfile(guildID, userID uint64) global.MSG {
+	if bot.Client.GuildService.FindGuild(guildID) == nil {
+		return Failed(100, "GUILD_NOT_FOUND")
+	}
+	profile, err := bot.Client.GuildService.FetchGuildMemberProfileInfo(guildID, userID)
+	if err != nil {
+		log.Warnf("获取频道 %v 成员 %v 资料时出现错误: %v", guildID, userID, err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	roles := make([]global.MSG, 0, len(profile.Roles))
+	for _, role := range profile.Roles {
+		roles = append(roles, global.MSG{
+			"role_id":   fU64(role.RoleId),
+			"role_name": role.RoleName,
+		})
+	}
+	return OK(global.MSG{
+		"tiny_id":    fU64(profile.TinyId),
+		"nickname":   profile.Nickname,
+		"avatar_url": profile.AvatarUrl,
+		"join_time":  profile.JoinTime,
+		"roles":      roles,
+	})
+}
+
+// CQGetGuildRoles 获取频道角色列表
+// @route(get_guild_roles)
+func (bot *CQBot) CQGetGuildRoles(guildID uint64) global.MSG {
+	r, err := bot.Client.GuildService.GetGuildRoles(guildID)
+	if err != nil {
+		log.Warnf("获取频道 %v 角色列表时出现错误: %v", guildID, err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	roles := make([]global.MSG, len(r))
+	for i, role := range r {
+		roles[i] = global.MSG{
+			"role_id":      fU64(role.RoleId),
+			"role_name":    role.RoleName,
+			"argb_color":   role.ArgbColor,
+			"independent":  role.Independent,
+			"member_count": role.Num,
+			"max_count":    role.MaxNum,
+			"owned":        role.Owned,
+			"disabled":     role.Disabled,
+		}
+	}
+	return OK(roles)
+}
+
+// CQCreateGuildRole 创建频道角色
+// @route(create_guild_role)
+func (bot *CQBot) CQCreateGuildRole(guildID uint64, name string, color uint32, independent bool, initialUsers gjson.Result) global.MSG {
+	userSlice := []uint64{}
+	if initialUsers.IsArray() {
+		for _, user := range initialUsers.Array() {
+			userSlice = append(userSlice, user.Uint())
+		}
+	}
+	role, err := bot.Client.GuildService.CreateGuildRole(guildID, name, color, independent, userSlice)
+	if err != nil {
+		log.Warnf("创建频道 %v 角色时出现错误: %v", guildID, err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	return OK(global.MSG{
+		"role_id": fU64(role),
+	})
+}
+
+// CQDeleteGuildRole 删除频道角色
+// @route(delete_guild_role)
+func (bot *CQBot) CQDeleteGuildRole(guildID uint64, roleID uint64) global.MSG {
+	err := bot.Client.GuildService.DeleteGuildRole(guildID, roleID)
+	if err != nil {
+		log.Warnf("删除频道 %v 角色时出现错误: %v", guildID, err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	return OK(nil)
+}
+
+// CQSetGuildMemberRole 设置用户在频道中的角色
+// @route(set_guild_member_role)
+func (bot *CQBot) CQSetGuildMemberRole(guildID uint64, set bool, roleID uint64, users gjson.Result) global.MSG {
+	userSlice := []uint64{}
+	if users.IsArray() {
+		for _, user := range users.Array() {
+			userSlice = append(userSlice, user.Uint())
+		}
+	}
+	err := bot.Client.GuildService.SetUserRoleInGuild(guildID, set, roleID, userSlice)
+	if err != nil {
+		log.Warnf("设置用户在频道 %v 中的角色时出现错误: %v", guildID, err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	return OK(nil)
+}
+
+// CQModifyRoleInGuild 修改频道角色
+// @route(update_guild_role)
+func (bot *CQBot) CQModifyRoleInGuild(guildID uint64, roleID uint64, name string, color uint32, indepedent bool) global.MSG {
+	err := bot.Client.GuildService.ModifyRoleInGuild(guildID, roleID, name, color, indepedent)
+	if err != nil {
+		log.Warnf("修改频道 %v 角色时出现错误: %v", guildID, err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	return OK(nil)
+}
+
+// CQGetTopicChannelFeeds 获取话题频道帖子列表
+// @route(get_topic_channel_feeds)
+func (bot *CQBot) CQGetTopicChannelFeeds(guildID, channelID uint64) global.MSG {
+	guild := bot.Client.GuildService.FindGuild(guildID)
+	if guild == nil {
+		return Failed(100, "GUILD_NOT_FOUND")
+	}
+	channel := guild.FindChannel(channelID)
+	if channel == nil {
+		return Failed(100, "CHANNEL_NOT_FOUND")
+	}
+	if channel.ChannelType != client.ChannelTypeTopic {
+		return Failed(100, "CHANNEL_TYPE_ERROR")
+	}
+	feeds, err := bot.Client.GuildService.GetTopicChannelFeeds(guildID, channelID)
+	if err != nil {
+		log.Warnf("获取频道 %v 帖子时出现错误: %v", channelID, err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	c := make([]global.MSG, 0, len(feeds))
+	for _, feed := range feeds {
+		c = append(c, convertChannelFeedInfo(feed))
+	}
+	return OK(c)
 }
 
 // CQGetFriendList 获取好友列表
 //
 // https://git.io/Jtz1L
-func (bot *CQBot) CQGetFriendList() MSG {
-	fs := make([]MSG, 0, len(bot.Client.FriendList))
+// @route(get_friend_list)
+func (bot *CQBot) CQGetFriendList() global.MSG {
+	fs := make([]global.MSG, 0, len(bot.Client.FriendList))
 	for _, f := range bot.Client.FriendList {
-		fs = append(fs, MSG{
+		fs = append(fs, global.MSG{
 			"nickname": f.Nickname,
 			"remark":   f.Remark,
 			"user_id":  f.Uin,
@@ -61,16 +339,73 @@ func (bot *CQBot) CQGetFriendList() MSG {
 	return OK(fs)
 }
 
+// CQGetUnidirectionalFriendList 获取单向好友列表
+//
+// @route(get_unidirectional_friend_list)
+func (bot *CQBot) CQGetUnidirectionalFriendList() global.MSG {
+	list, err := bot.Client.GetUnidirectionalFriendList()
+	if err != nil {
+		log.Warnf("获取单向好友列表时出现错误: %v", err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	fs := make([]global.MSG, 0, len(list))
+	for _, f := range list {
+		fs = append(fs, global.MSG{
+			"nickname": f.Nickname,
+			"user_id":  f.Uin,
+			"source":   f.Source,
+		})
+	}
+	return OK(fs)
+}
+
+// CQDeleteUnidirectionalFriend 删除单向好友
+//
+// @route(delete_unidirectional_friend)
+// @rename(uin->user_id)
+func (bot *CQBot) CQDeleteUnidirectionalFriend(uin int64) global.MSG {
+	list, err := bot.Client.GetUnidirectionalFriendList()
+	if err != nil {
+		log.Warnf("获取单向好友列表时出现错误: %v", err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	for _, f := range list {
+		if f.Uin == uin {
+			if err = bot.Client.DeleteUnidirectionalFriend(uin); err != nil {
+				log.Warnf("删除单向好友时出现错误: %v", err)
+				return Failed(100, "API_ERROR", err.Error())
+			}
+			return OK(nil)
+		}
+	}
+	return Failed(100, "FRIEND_NOT_FOUND", "好友不存在")
+}
+
+// CQDeleteFriend 删除好友
+// @route(delete_friend)
+// @rename(uin->"[user_id\x2Cid].0")
+func (bot *CQBot) CQDeleteFriend(uin int64) global.MSG {
+	if bot.Client.FindFriend(uin) == nil {
+		return Failed(100, "FRIEND_NOT_FOUND", "好友不存在")
+	}
+	if err := bot.Client.DeleteFriend(uin); err != nil {
+		log.Warnf("删除好友时出现错误: %v", err)
+		return Failed(100, "DELETE_API_ERROR", err.Error())
+	}
+	return OK(nil)
+}
+
 // CQGetGroupList 获取群列表
 //
 // https://git.io/Jtz1t
-func (bot *CQBot) CQGetGroupList(noCache bool) MSG {
-	gs := make([]MSG, 0, len(bot.Client.GroupList))
+// @route(get_group_list)
+func (bot *CQBot) CQGetGroupList(noCache bool) global.MSG {
+	gs := make([]global.MSG, 0, len(bot.Client.GroupList))
 	if noCache {
 		_ = bot.Client.ReloadGroupList()
 	}
 	for _, g := range bot.Client.GroupList {
-		gs = append(gs, MSG{
+		gs = append(gs, global.MSG{
 			"group_id":          g.Code,
 			"group_name":        g.Name,
 			"group_memo":        g.Memo,
@@ -86,7 +421,8 @@ func (bot *CQBot) CQGetGroupList(noCache bool) MSG {
 // CQGetGroupInfo 获取群信息
 //
 // https://git.io/Jtz1O
-func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool) MSG {
+// @route(get_group_info)
+func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool) global.MSG {
 	group := bot.Client.FindGroup(groupID)
 	if group == nil || noCache {
 		group, _ = bot.Client.GetGroupInfo(groupID)
@@ -99,7 +435,7 @@ func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool) MSG {
 		}
 		for _, g := range info {
 			if g.Code == groupID {
-				return OK(MSG{
+				return OK(global.MSG{
 					"group_id":          g.Code,
 					"group_name":        g.Name,
 					"group_memo":        g.Memo,
@@ -111,7 +447,7 @@ func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool) MSG {
 			}
 		}
 	} else {
-		return OK(MSG{
+		return OK(global.MSG{
 			"group_id":          group.Code,
 			"group_name":        group.Name,
 			"group_memo":        group.Memo,
@@ -127,7 +463,8 @@ func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool) MSG {
 // CQGetGroupMemberList 获取群成员列表
 //
 // https://git.io/Jtz13
-func (bot *CQBot) CQGetGroupMemberList(groupID int64, noCache bool) MSG {
+// @route(get_group_member_list)
+func (bot *CQBot) CQGetGroupMemberList(groupID int64, noCache bool) global.MSG {
 	group := bot.Client.FindGroup(groupID)
 	if group == nil {
 		return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
@@ -140,7 +477,7 @@ func (bot *CQBot) CQGetGroupMemberList(groupID int64, noCache bool) MSG {
 		}
 		group.Members = t
 	}
-	members := make([]MSG, 0, len(group.Members))
+	members := make([]global.MSG, 0, len(group.Members))
 	for _, m := range group.Members {
 		members = append(members, convertGroupMemberInfo(groupID, m))
 	}
@@ -150,7 +487,8 @@ func (bot *CQBot) CQGetGroupMemberList(groupID int64, noCache bool) MSG {
 // CQGetGroupMemberInfo 获取群成员信息
 //
 // https://git.io/Jtz1s
-func (bot *CQBot) CQGetGroupMemberInfo(groupID, userID int64, noCache bool) MSG {
+// @route(get_group_member_info)
+func (bot *CQBot) CQGetGroupMemberInfo(groupID, userID int64, noCache bool) global.MSG {
 	group := bot.Client.FindGroup(groupID)
 	if group == nil {
 		return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
@@ -175,10 +513,11 @@ func (bot *CQBot) CQGetGroupMemberInfo(groupID, userID int64, noCache bool) MSG 
 // CQGetGroupFileSystemInfo 扩展API-获取群文件系统信息
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4%E6%96%87%E4%BB%B6%E7%B3%BB%E7%BB%9F%E4%BF%A1%E6%81%AF
-func (bot *CQBot) CQGetGroupFileSystemInfo(groupID int64) MSG {
+// @route(get_group_file_system_info)
+func (bot *CQBot) CQGetGroupFileSystemInfo(groupID int64) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
-		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
+		log.Warnf("获取群 %v 文件系统信息失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	return OK(fs)
@@ -187,18 +526,19 @@ func (bot *CQBot) CQGetGroupFileSystemInfo(groupID int64) MSG {
 // CQGetGroupRootFiles 扩展API-获取群根目录文件列表
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4%E6%A0%B9%E7%9B%AE%E5%BD%95%E6%96%87%E4%BB%B6%E5%88%97%E8%A1%A8
-func (bot *CQBot) CQGetGroupRootFiles(groupID int64) MSG {
+// @route(get_group_root_files)
+func (bot *CQBot) CQGetGroupRootFiles(groupID int64) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
-		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
+		log.Warnf("获取群 %v 文件系统信息失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	files, folders, err := fs.Root()
 	if err != nil {
-		log.Errorf("获取群 %v 根目录文件失败: %v", groupID, err)
+		log.Warnf("获取群 %v 根目录文件失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"files":   files,
 		"folders": folders,
 	})
@@ -207,18 +547,19 @@ func (bot *CQBot) CQGetGroupRootFiles(groupID int64) MSG {
 // CQGetGroupFilesByFolderID 扩展API-获取群子目录文件列表
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4%E5%AD%90%E7%9B%AE%E5%BD%95%E6%96%87%E4%BB%B6%E5%88%97%E8%A1%A8
-func (bot *CQBot) CQGetGroupFilesByFolderID(groupID int64, folderID string) MSG {
+// @route(get_group_files_by_folder)
+func (bot *CQBot) CQGetGroupFilesByFolderID(groupID int64, folderID string) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
-		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
+		log.Warnf("获取群 %v 文件系统信息失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	files, folders, err := fs.GetFilesByFolder(folderID)
 	if err != nil {
-		log.Errorf("获取群 %v 根目录 %v 子文件失败: %v", groupID, folderID, err)
+		log.Warnf("获取群 %v 根目录 %v 子文件失败: %v", groupID, folderID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"files":   files,
 		"folders": folders,
 	})
@@ -227,12 +568,14 @@ func (bot *CQBot) CQGetGroupFilesByFolderID(groupID int64, folderID string) MSG 
 // CQGetGroupFileURL 扩展API-获取群文件资源链接
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4%E6%96%87%E4%BB%B6%E8%B5%84%E6%BA%90%E9%93%BE%E6%8E%A5
-func (bot *CQBot) CQGetGroupFileURL(groupID int64, fileID string, busID int32) MSG {
+// @route(get_group_file_url)
+// @rename(bus_id->"[busid\x2Cbus_id].0")
+func (bot *CQBot) CQGetGroupFileURL(groupID int64, fileID string, busID int32) global.MSG {
 	url := bot.Client.GetGroupFileUrl(groupID, fileID, busID)
 	if url == "" {
 		return Failed(100, "FILE_SYSTEM_API_ERROR")
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"url": url,
 	})
 }
@@ -240,22 +583,73 @@ func (bot *CQBot) CQGetGroupFileURL(groupID int64, fileID string, busID int32) M
 // CQUploadGroupFile 扩展API-上传群文件
 //
 // https://docs.go-cqhttp.org/api/#%E4%B8%8A%E4%BC%A0%E7%BE%A4%E6%96%87%E4%BB%B6
-func (bot *CQBot) CQUploadGroupFile(groupID int64, file, name, folder string) MSG {
+// @route(upload_group_file)
+func (bot *CQBot) CQUploadGroupFile(groupID int64, file, name, folder string) global.MSG {
 	if !global.PathExists(file) {
-		log.Errorf("上传群文件 %v 失败: 文件不存在", file)
+		log.Warnf("上传群文件 %v 失败: 文件不存在", file)
 		return Failed(100, "FILE_NOT_FOUND", "文件不存在")
 	}
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
-		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
+		log.Warnf("获取群 %v 文件系统信息失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	if folder == "" {
 		folder = "/"
 	}
 	if err = fs.UploadFile(file, name, folder); err != nil {
-		log.Errorf("上传群 %v 文件 %v 失败: %v", groupID, file, err)
+		log.Warnf("上传群 %v 文件 %v 失败: %v", groupID, file, err)
 		return Failed(100, "FILE_SYSTEM_UPLOAD_API_ERROR", err.Error())
+	}
+	return OK(nil)
+}
+
+// CQGroupFileCreateFolder 拓展API-创建群文件文件夹
+//
+// @route(create_group_file_folder)
+func (bot *CQBot) CQGroupFileCreateFolder(groupID int64, parentID, name string) global.MSG {
+	fs, err := bot.Client.GetGroupFileSystem(groupID)
+	if err != nil {
+		log.Warnf("获取群 %v 文件系统信息失败: %v", groupID, err)
+		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
+	}
+	if err = fs.CreateFolder(parentID, name); err != nil {
+		log.Warnf("创建群 %v 文件夹失败: %v", groupID, err)
+		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
+	}
+	return OK(nil)
+}
+
+// CQGroupFileDeleteFolder 拓展API-删除群文件文件夹
+//
+// @route(delete_group_folder)
+// @rename(id->folder_id)
+func (bot *CQBot) CQGroupFileDeleteFolder(groupID int64, id string) global.MSG {
+	fs, err := bot.Client.GetGroupFileSystem(groupID)
+	if err != nil {
+		log.Warnf("获取群 %v 文件系统信息失败: %v", groupID, err)
+		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
+	}
+	if err = fs.DeleteFolder(id); err != nil {
+		log.Warnf("删除群 %v 文件夹 %v 时出现文件: %v", groupID, id, err)
+		return Failed(200, "FILE_SYSTEM_API_ERROR", err.Error())
+	}
+	return OK(nil)
+}
+
+// CQGroupFileDeleteFile 拓展API-删除群文件
+//
+// @route(delete_group_file)
+// @rename(id->file_id, bus_id->"[busid\x2Cbus_id].0")
+func (bot *CQBot) CQGroupFileDeleteFile(groupID int64, id string, busID int32) global.MSG {
+	fs, err := bot.Client.GetGroupFileSystem(groupID)
+	if err != nil {
+		log.Warnf("获取群 %v 文件系统信息失败: %v", groupID, err)
+		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
+	}
+	if res := fs.DeleteFile("", id, busID); res != "" {
+		log.Warnf("删除群 %v 文件 %v 时出现文件: %v", groupID, id, res)
+		return Failed(200, "FILE_SYSTEM_API_ERROR", res)
 	}
 	return OK(nil)
 }
@@ -263,7 +657,8 @@ func (bot *CQBot) CQUploadGroupFile(groupID int64, file, name, folder string) MS
 // CQGetWordSlices 隐藏API-获取中文分词
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E4%B8%AD%E6%96%87%E5%88%86%E8%AF%8D-%E9%9A%90%E8%97%8F-api
-func (bot *CQBot) CQGetWordSlices(content string) MSG {
+// @route(.get_word_slices)
+func (bot *CQBot) CQGetWordSlices(content string) global.MSG {
 	slices, err := bot.Client.GetWordSegmentation(content)
 	if err != nil {
 		return Failed(100, "WORD_SEGMENTATION_API_ERROR", err.Error())
@@ -271,14 +666,33 @@ func (bot *CQBot) CQGetWordSlices(content string) MSG {
 	for i := 0; i < len(slices); i++ {
 		slices[i] = strings.ReplaceAll(slices[i], "\u0000", "")
 	}
-	return OK(MSG{"slices": slices})
+	return OK(global.MSG{"slices": slices})
+}
+
+// CQSendMessage 发送消息
+//
+// @route(send_msg)
+// @rename(m->message)
+func (bot *CQBot) CQSendMessage(groupID, userID int64, m gjson.Result, messageType string, autoEscape bool) global.MSG {
+	switch {
+	case messageType == "group":
+		return bot.CQSendGroupMessage(groupID, m, autoEscape)
+	case messageType == "private":
+		fallthrough
+	case userID != 0:
+		return bot.CQSendPrivateMessage(userID, groupID, m, autoEscape)
+	case groupID != 0:
+		return bot.CQSendGroupMessage(groupID, m, autoEscape)
+	}
+	return global.MSG{}
 }
 
 // CQSendGroupMessage 发送群消息
 //
 // https://git.io/Jtz1c
-func (bot *CQBot) CQSendGroupMessage(groupID int64, i interface{}, autoEscape bool) MSG {
-	var str string
+// @route(send_group_msg)
+// @rename(m->message)
+func (bot *CQBot) CQSendGroupMessage(groupID int64, m gjson.Result, autoEscape bool) global.MSG {
 	group := bot.Client.FindGroup(groupID)
 	if group == nil {
 		return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
@@ -286,230 +700,263 @@ func (bot *CQBot) CQSendGroupMessage(groupID int64, i interface{}, autoEscape bo
 	fixAt := func(elem []message.IMessageElement) {
 		for _, e := range elem {
 			if at, ok := e.(*message.AtElement); ok && at.Target != 0 && at.Display == "" {
-				at.Display = "@" + func() string {
-					mem := group.FindMember(at.Target)
-					if mem != nil {
-						return mem.DisplayName()
-					}
-					return strconv.FormatInt(at.Target, 10)
-				}()
+				mem := group.FindMember(at.Target)
+				if mem != nil {
+					at.Display = "@" + mem.DisplayName()
+				} else {
+					at.Display = "@" + strconv.FormatInt(at.Target, 10)
+				}
 			}
 		}
 	}
-	if m, ok := i.(gjson.Result); ok {
-		if m.Type == gjson.JSON {
-			elem := bot.ConvertObjectMessage(m, true)
-			fixAt(elem)
-			mid := bot.SendGroupMessage(groupID, &message.SendingMessage{Elements: elem})
-			if mid == -1 {
-				return Failed(100, "SEND_MSG_API_ERROR", "请参考输出")
-			}
-			log.Infof("发送群 %v(%v) 的消息: %v (%v)", group.Name, groupID, limitedString(ToStringMessage(elem, int64(mid))), mid)
-			return OK(MSG{"message_id": mid})
-		}
-		str = func() string {
-			if m.Str != "" {
-				return m.Str
-			}
-			return m.Raw
-		}()
-	} else if s, ok := i.(string); ok {
-		str = s
-	}
-	if str == "" {
-		log.Warnf("群消息发送失败: 信息为空. MSG: %v", i)
-		return Failed(100, "EMPTY_MSG_ERROR", "消息为空")
-	}
+
 	var elem []message.IMessageElement
-	if autoEscape {
-		elem = append(elem, message.NewText(str))
+	if m.Type == gjson.JSON {
+		elem = bot.ConvertObjectMessage(m, MessageSourceGroup)
 	} else {
-		elem = bot.ConvertStringMessage(str, true)
+		str := m.String()
+		if str == "" {
+			log.Warn("群消息发送失败: 信息为空.")
+			return Failed(100, "EMPTY_MSG_ERROR", "消息为空")
+		}
+		if autoEscape {
+			elem = []message.IMessageElement{message.NewText(str)}
+		} else {
+			elem = bot.ConvertStringMessage(str, MessageSourceGroup)
+		}
 	}
 	fixAt(elem)
 	mid := bot.SendGroupMessage(groupID, &message.SendingMessage{Elements: elem})
 	if mid == -1 {
-		return Failed(100, "SEND_MSG_API_ERROR", "请参考输出")
+		return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
 	}
-	log.Infof("发送群 %v(%v) 的消息: %v (%v)", group.Name, groupID, limitedString(str), mid)
-	return OK(MSG{"message_id": mid})
+	log.Infof("发送群 %v(%v) 的消息: %v (%v)", group.Name, groupID, limitedString(m.String()), mid)
+	return OK(global.MSG{"message_id": mid})
 }
 
-// CQSendGroupForwardMessage 扩展API-发送合并转发(群)
+// CQSendGuildChannelMessage 发送频道消息
 //
-// https://docs.go-cqhttp.org/api/#%E5%8F%91%E9%80%81%E5%90%88%E5%B9%B6%E8%BD%AC%E5%8F%91-%E7%BE%A4
-func (bot *CQBot) CQSendGroupForwardMessage(groupID int64, m gjson.Result) MSG {
-	if m.Type != gjson.JSON {
-		return Failed(100)
+// @route(send_guild_channel_msg)
+// @rename(m->message)
+func (bot *CQBot) CQSendGuildChannelMessage(guildID, channelID uint64, m gjson.Result, autoEscape bool) global.MSG {
+	guild := bot.Client.GuildService.FindGuild(guildID)
+	if guild == nil {
+		return Failed(100, "GUILD_NOT_FOUND", "频道不存在")
 	}
-	var sendNodes []*message.ForwardNode
-	ts := time.Now().Add(-time.Minute * 5)
-	hasCustom := func() bool {
-		for _, item := range m.Array() {
-			if item.Get("data.uin").Exists() {
-				return true
+	channel := guild.FindChannel(channelID)
+	if channel == nil {
+		return Failed(100, "CHANNEL_NOT_FOUND", "子频道不存在")
+	}
+	if channel.ChannelType != client.ChannelTypeText {
+		log.Warnf("无法发送频道信息: 频道类型错误, 不接受文本信息")
+		return Failed(100, "CHANNEL_NOT_SUPPORTED_TEXT_MSG", "子频道类型错误, 无法发送文本信息")
+	}
+	fixAt := func(elem []message.IMessageElement) {
+		for _, e := range elem {
+			if at, ok := e.(*message.AtElement); ok && at.Target != 0 && at.Display == "" {
+				mem, _ := bot.Client.GuildService.FetchGuildMemberProfileInfo(guildID, uint64(at.Target))
+				if mem != nil {
+					at.Display = "@" + mem.Nickname
+				} else {
+					at.Display = "@" + strconv.FormatInt(at.Target, 10)
+				}
 			}
 		}
-		return false
-	}()
-	var convert func(e gjson.Result) []*message.ForwardNode
-	convert = func(e gjson.Result) (nodes []*message.ForwardNode) {
+	}
+
+	var elem []message.IMessageElement
+	if m.Type == gjson.JSON {
+		elem = bot.ConvertObjectMessage(m, MessageSourceGuildChannel)
+	} else {
+		str := m.String()
+		if str == "" {
+			log.Warn("频道发送失败: 信息为空.")
+			return Failed(100, "EMPTY_MSG_ERROR", "消息为空")
+		}
+		if autoEscape {
+			elem = []message.IMessageElement{message.NewText(str)}
+		} else {
+			elem = bot.ConvertStringMessage(str, MessageSourceGuildChannel)
+		}
+	}
+	fixAt(elem)
+	mid := bot.SendGuildChannelMessage(guildID, channelID, &message.SendingMessage{Elements: elem})
+	if mid == "" {
+		return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
+	}
+	log.Infof("发送频道 %v(%v) 子频道 %v(%v) 的消息: %v (%v)", guild.GuildName, guild.GuildId, channel.ChannelName, channel.ChannelId, limitedString(m.String()), mid)
+	return OK(global.MSG{"message_id": mid})
+}
+
+func (bot *CQBot) uploadForwardElement(m gjson.Result, groupID int64) *message.ForwardElement {
+	ts := time.Now().Add(-time.Minute * 5)
+	fm := message.NewForwardMessage()
+
+	var lazyUpload []func()
+	var wg sync.WaitGroup
+	resolveElement := func(elems []message.IMessageElement) []message.IMessageElement {
+		for i, elem := range elems {
+			iescape := i
+			switch o := elem.(type) {
+			case *LocalImageElement, *LocalVideoElement:
+				wg.Add(1)
+				lazyUpload = append(lazyUpload, func() {
+					defer wg.Done()
+					gm, err := bot.uploadMedia(o, groupID, true)
+					if err != nil {
+						log.Warnf("警告: 群 %d %s上传失败: %v", groupID, o.Type().String(), err)
+					} else {
+						elems[iescape] = gm
+					}
+				})
+			}
+		}
+		return elems
+	}
+
+	convert := func(e gjson.Result) *message.ForwardNode {
 		if e.Get("type").Str != "node" {
 			return nil
 		}
 		ts.Add(time.Second)
 		if e.Get("data.id").Exists() {
-			i, _ := strconv.Atoi(e.Get("data.id").String())
-			m := bot.GetMessage(int32(i))
+			i := e.Get("data.id").Int()
+			m, _ := db.GetGroupMessageByGlobalID(int32(i))
 			if m != nil {
-				sender := m["sender"].(message.Sender)
-				nodes = append(nodes, &message.ForwardNode{
-					SenderId:   sender.Uin,
-					SenderName: (&sender).DisplayName(),
+				return &message.ForwardNode{
+					SenderId:   m.Attribute.SenderUin,
+					SenderName: m.Attribute.SenderName,
 					Time: func() int32 {
-						msgTime := m["time"].(int32)
-						if hasCustom && msgTime == 0 {
+						msgTime := m.Attribute.Timestamp
+						if msgTime == 0 {
 							return int32(ts.Unix())
 						}
-						return msgTime
+						return int32(msgTime)
 					}(),
-					Message: bot.ConvertStringMessage(m["message"].(string), true),
-				})
-				return
+					Message: resolveElement(bot.ConvertContentMessage(m.Content, MessageSourceGroup)),
+				}
 			}
 			log.Warnf("警告: 引用消息 %v 错误或数据库未开启.", e.Get("data.id").Str)
-			return
+			return nil
 		}
-		uin, _ := strconv.ParseInt(e.Get("data.uin").Str, 10, 64)
-		msgTime, err := strconv.ParseInt(e.Get("data.time").Str, 10, 64)
-		if err != nil {
+		uin := e.Get("data.[user_id,uin].0").Int()
+		msgTime := e.Get("data.time").Int()
+		if msgTime == 0 {
 			msgTime = ts.Unix()
 		}
 		name := e.Get("data.name").Str
 		c := e.Get("data.content")
 		if c.IsArray() {
-			flag := false
+			nested := false
 			c.ForEach(func(_, value gjson.Result) bool {
-				if value.Get("type").String() == "node" {
-					flag = true
+				if value.Get("type").Str == "node" {
+					nested = true
 					return false
 				}
 				return true
 			})
-			if flag {
-				var taowa []*message.ForwardNode
-				for _, item := range c.Array() {
-					taowa = append(taowa, convert(item)...)
-				}
-				nodes = append(nodes, &message.ForwardNode{
+			if nested { // 处理嵌套
+				fe := bot.uploadForwardElement(c, groupID)
+				return &message.ForwardNode{
 					SenderId:   uin,
 					SenderName: name,
 					Time:       int32(msgTime),
-					Message:    []message.IMessageElement{bot.Client.UploadGroupForwardMessage(groupID, &message.ForwardMessage{Nodes: taowa})},
-				})
-				return
+					Message:    []message.IMessageElement{fe},
+				}
 			}
 		}
-		content := bot.ConvertObjectMessage(e.Get("data.content"), true)
+		content := bot.ConvertObjectMessage(c, MessageSourceGroup)
 		if uin != 0 && name != "" && len(content) > 0 {
-			var newElem []message.IMessageElement
-			for _, elem := range content {
-				if img, ok := elem.(*LocalImageElement); ok {
-					gm, err := bot.UploadLocalImageAsGroup(groupID, img)
-					if err != nil {
-						log.Warnf("警告：群 %v 图片上传失败: %v", groupID, err)
-						continue
-					}
-					newElem = append(newElem, gm)
-					continue
-				}
-				if video, ok := elem.(*LocalVideoElement); ok {
-					gm, err := bot.UploadLocalVideo(groupID, video)
-					if err != nil {
-						log.Warnf("警告：群 %v 视频上传失败: %v", groupID, err)
-						continue
-					}
-					newElem = append(newElem, gm)
-					continue
-				}
-				newElem = append(newElem, elem)
-			}
-			nodes = append(nodes, &message.ForwardNode{
+			return &message.ForwardNode{
 				SenderId:   uin,
 				SenderName: name,
 				Time:       int32(msgTime),
-				Message:    newElem,
-			})
-			return
+				Message:    resolveElement(content),
+			}
 		}
-		log.Warnf("警告: 非法 Forward node 将跳过")
-		return
+		log.Warnf("警告: 非法 Forward node 将跳过. uin: %v name: %v content count: %v", uin, name, len(content))
+		return nil
 	}
+
 	if m.IsArray() {
 		for _, item := range m.Array() {
-			sendNodes = append(sendNodes, convert(item)...)
+			node := convert(item)
+			if node != nil {
+				fm.AddNode(node)
+			}
 		}
 	} else {
-		sendNodes = convert(m)
+		node := convert(m)
+		if node != nil {
+			fm.AddNode(node)
+		}
 	}
-	if len(sendNodes) > 0 {
-		ret := bot.Client.SendGroupForwardMessage(groupID, &message.ForwardMessage{Nodes: sendNodes})
+
+	for _, upload := range lazyUpload {
+		go upload()
+	}
+	wg.Wait()
+
+	return bot.Client.UploadGroupForwardMessage(groupID, fm)
+}
+
+// CQSendGroupForwardMessage 扩展API-发送合并转发(群)
+//
+// https://docs.go-cqhttp.org/api/#%E5%8F%91%E9%80%81%E5%90%88%E5%B9%B6%E8%BD%AC%E5%8F%91-%E7%BE%A4
+// @route(send_group_forward_msg)
+// @rename(m->messages)
+func (bot *CQBot) CQSendGroupForwardMessage(groupID int64, m gjson.Result) global.MSG {
+	if m.Type != gjson.JSON {
+		return Failed(100)
+	}
+
+	fe := bot.uploadForwardElement(m, groupID)
+	if fe != nil {
+		ret := bot.Client.SendGroupForwardMessage(groupID, fe)
 		if ret == nil || ret.Id == -1 {
 			log.Warnf("合并转发(群)消息发送失败: 账号可能被风控.")
-			return Failed(100, "SEND_MSG_API_ERROR", "请参考输出")
+			return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
 		}
-		return OK(MSG{
+		return OK(global.MSG{
 			"message_id": bot.InsertGroupMessage(ret),
 		})
 	}
-	return Failed(100)
+	return Failed(100, "EMPTY_NODES", "未找到任何可发送的合并转发信息")
 }
 
 // CQSendPrivateMessage 发送私聊消息
 //
 // https://git.io/Jtz1l
-func (bot *CQBot) CQSendPrivateMessage(userID int64, groupID int64, i interface{}, autoEscape bool) MSG {
-	var str string
-	if m, ok := i.(gjson.Result); ok {
-		if m.Type == gjson.JSON {
-			elem := bot.ConvertObjectMessage(m, false)
-			mid := bot.SendPrivateMessage(userID, groupID, &message.SendingMessage{Elements: elem})
-			if mid == -1 {
-				return Failed(100, "SEND_MSG_API_ERROR", "请参考输出")
-			}
-			log.Infof("发送好友 %v(%v)  的消息: %v (%v)", userID, userID, limitedString(m.String()), mid)
-			return OK(MSG{"message_id": mid})
-		}
-		str = func() string {
-			if m.Str != "" {
-				return m.Str
-			}
-			return m.Raw
-		}()
-	} else if s, ok := i.(string); ok {
-		str = s
-	}
-	if str == "" {
-		return Failed(100, "EMPTY_MSG_ERROR", "消息为空")
-	}
+// @route(send_private_msg)
+// @rename(m->message)
+func (bot *CQBot) CQSendPrivateMessage(userID int64, groupID int64, m gjson.Result, autoEscape bool) global.MSG {
 	var elem []message.IMessageElement
-	if autoEscape {
-		elem = append(elem, message.NewText(str))
+	if m.Type == gjson.JSON {
+		elem = bot.ConvertObjectMessage(m, MessageSourcePrivate)
 	} else {
-		elem = bot.ConvertStringMessage(str, false)
+		str := m.String()
+		if str == "" {
+			return Failed(100, "EMPTY_MSG_ERROR", "消息为空")
+		}
+		if autoEscape {
+			elem = []message.IMessageElement{message.NewText(str)}
+		} else {
+			elem = bot.ConvertStringMessage(str, MessageSourcePrivate)
+		}
 	}
 	mid := bot.SendPrivateMessage(userID, groupID, &message.SendingMessage{Elements: elem})
 	if mid == -1 {
-		return Failed(100, "SEND_MSG_API_ERROR", "请参考输出")
+		return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
 	}
-	log.Infof("发送好友 %v(%v)  的消息: %v (%v)", userID, userID, limitedString(str), mid)
-	return OK(MSG{"message_id": mid})
+	log.Infof("发送好友 %v(%v)  的消息: %v (%v)", userID, userID, limitedString(m.String()), mid)
+	return OK(global.MSG{"message_id": mid})
 }
 
 // CQSetGroupCard 设置群名片(群备注)
 //
 // https://git.io/Jtz1B
-func (bot *CQBot) CQSetGroupCard(groupID, userID int64, card string) MSG {
+// @route(set_group_card)
+func (bot *CQBot) CQSetGroupCard(groupID, userID int64, card string) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		if m := g.FindMember(userID); m != nil {
 			m.EditCard(card)
@@ -522,7 +969,9 @@ func (bot *CQBot) CQSetGroupCard(groupID, userID int64, card string) MSG {
 // CQSetGroupSpecialTitle 设置群组专属头衔
 //
 // https://git.io/Jtz10
-func (bot *CQBot) CQSetGroupSpecialTitle(groupID, userID int64, title string) MSG {
+// @route(set_group_special_title)
+// @rename(title->special_title)
+func (bot *CQBot) CQSetGroupSpecialTitle(groupID, userID int64, title string) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		if m := g.FindMember(userID); m != nil {
 			m.EditSpecialTitle(title)
@@ -535,7 +984,9 @@ func (bot *CQBot) CQSetGroupSpecialTitle(groupID, userID int64, title string) MS
 // CQSetGroupName 设置群名
 //
 // https://git.io/Jtz12
-func (bot *CQBot) CQSetGroupName(groupID int64, name string) MSG {
+// @route(set_group_name)
+// @rename(name->group_name)
+func (bot *CQBot) CQSetGroupName(groupID int64, name string) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		g.UpdateName(name)
 		return OK(nil)
@@ -546,12 +997,28 @@ func (bot *CQBot) CQSetGroupName(groupID int64, name string) MSG {
 // CQSetGroupMemo 扩展API-发送群公告
 //
 // https://docs.go-cqhttp.org/api/#%E5%8F%91%E9%80%81%E7%BE%A4%E5%85%AC%E5%91%8A
-func (bot *CQBot) CQSetGroupMemo(groupID int64, msg string) MSG {
+// @route(_send_group_notice)
+// @rename(msg->content, img->image)
+func (bot *CQBot) CQSetGroupMemo(groupID int64, msg, img string) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		if g.SelfPermission() == client.Member {
 			return Failed(100, "PERMISSION_DENIED", "权限不足")
 		}
-		_ = bot.Client.AddGroupNoticeSimple(groupID, msg)
+		if img != "" {
+			data, err := global.FindFile(img, "", global.ImagePath)
+			if err != nil {
+				return Failed(100, "IMAGE_NOT_FOUND", "图片未找到")
+			}
+			err = bot.Client.AddGroupNoticeWithPic(groupID, msg, data)
+			if err != nil {
+				return Failed(100, "SEND_NOTICE_ERROR", err.Error())
+			}
+		} else {
+			err := bot.Client.AddGroupNoticeSimple(groupID, msg)
+			if err != nil {
+				return Failed(100, "SEND_NOTICE_ERROR", err.Error())
+			}
+		}
 		return OK(nil)
 	}
 	return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
@@ -560,7 +1027,9 @@ func (bot *CQBot) CQSetGroupMemo(groupID int64, msg string) MSG {
 // CQSetGroupKick 群组踢人
 //
 // https://git.io/Jtz1V
-func (bot *CQBot) CQSetGroupKick(groupID, userID int64, msg string, block bool) MSG {
+// @route(set_group_kick)
+// @rename(msg->message, block->reject_add_request)
+func (bot *CQBot) CQSetGroupKick(groupID int64, userID int64, msg string, block bool) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		if m := g.FindMember(userID); m != nil {
 			err := m.Kick(msg, block)
@@ -576,7 +1045,9 @@ func (bot *CQBot) CQSetGroupKick(groupID, userID int64, msg string, block bool) 
 // CQSetGroupBan 群组单人禁言
 //
 // https://git.io/Jtz1w
-func (bot *CQBot) CQSetGroupBan(groupID, userID int64, duration uint32) MSG {
+// @route(set_group_ban)
+// @default(duration=1800)
+func (bot *CQBot) CQSetGroupBan(groupID, userID int64, duration uint32) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		if m := g.FindMember(userID); m != nil {
 			err := m.Mute(duration)
@@ -595,7 +1066,9 @@ func (bot *CQBot) CQSetGroupBan(groupID, userID int64, duration uint32) MSG {
 // CQSetGroupWholeBan 群组全员禁言
 //
 // https://git.io/Jtz1o
-func (bot *CQBot) CQSetGroupWholeBan(groupID int64, enable bool) MSG {
+// @route(set_group_whole_ban)
+// @default(enable=true)
+func (bot *CQBot) CQSetGroupWholeBan(groupID int64, enable bool) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		g.MuteAll(enable)
 		return OK(nil)
@@ -606,7 +1079,8 @@ func (bot *CQBot) CQSetGroupWholeBan(groupID int64, enable bool) MSG {
 // CQSetGroupLeave 退出群组
 //
 // https://git.io/Jtz1K
-func (bot *CQBot) CQSetGroupLeave(groupID int64) MSG {
+// @route(set_group_leave)
+func (bot *CQBot) CQSetGroupLeave(groupID int64) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		g.Quit()
 		return OK(nil)
@@ -617,7 +1091,8 @@ func (bot *CQBot) CQSetGroupLeave(groupID int64) MSG {
 // CQGetAtAllRemain 扩展API-获取群 @全体成员 剩余次数
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4-%E5%85%A8%E4%BD%93%E6%88%90%E5%91%98-%E5%89%A9%E4%BD%99%E6%AC%A1%E6%95%B0
-func (bot *CQBot) CQGetAtAllRemain(groupID int64) MSG {
+// @route(get_group_at_all_remain)
+func (bot *CQBot) CQGetAtAllRemain(groupID int64) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		i, err := bot.Client.GetAtAllRemain(groupID)
 		if err != nil {
@@ -631,7 +1106,9 @@ func (bot *CQBot) CQGetAtAllRemain(groupID int64) MSG {
 // CQProcessFriendRequest 处理加好友请求
 //
 // https://git.io/Jtz11
-func (bot *CQBot) CQProcessFriendRequest(flag string, approve bool) MSG {
+// @route(set_friend_add_request)
+// @default(approve=true)
+func (bot *CQBot) CQProcessFriendRequest(flag string, approve bool) global.MSG {
 	req, ok := bot.friendReqCache.Load(flag)
 	if !ok {
 		return Failed(100, "FLAG_NOT_FOUND", "FLAG不存在")
@@ -647,17 +1124,20 @@ func (bot *CQBot) CQProcessFriendRequest(flag string, approve bool) MSG {
 // CQProcessGroupRequest 处理加群请求／邀请
 //
 // https://git.io/Jtz1D
-func (bot *CQBot) CQProcessGroupRequest(flag, subType, reason string, approve bool) MSG {
+// @route(set_group_add_request)
+// @rename(sub_type->"[sub_type\x2Ctype].0")
+// @default(approve=true)
+func (bot *CQBot) CQProcessGroupRequest(flag, subType, reason string, approve bool) global.MSG {
 	msgs, err := bot.Client.GetGroupSystemMessages()
 	if err != nil {
-		log.Errorf("获取群系统消息失败: %v", err)
+		log.Warnf("获取群系统消息失败: %v", err)
 		return Failed(100, "SYSTEM_MSG_API_ERROR", err.Error())
 	}
 	if subType == "add" {
 		for _, req := range msgs.JoinRequests {
 			if strconv.FormatInt(req.RequestId, 10) == flag {
 				if req.Checked {
-					log.Errorf("处理群系统消息失败: 无法操作已处理的消息.")
+					log.Warnf("处理群系统消息失败: 无法操作已处理的消息.")
 					return Failed(100, "FLAG_HAS_BEEN_CHECKED", "消息已被处理")
 				}
 				if approve {
@@ -672,7 +1152,7 @@ func (bot *CQBot) CQProcessGroupRequest(flag, subType, reason string, approve bo
 		for _, req := range msgs.InvitedRequests {
 			if strconv.FormatInt(req.RequestId, 10) == flag {
 				if req.Checked {
-					log.Errorf("处理群系统消息失败: 无法操作已处理的消息.")
+					log.Warnf("处理群系统消息失败: 无法操作已处理的消息.")
 					return Failed(100, "FLAG_HAS_BEEN_CHECKED", "消息已被处理")
 				}
 				if approve {
@@ -684,37 +1164,37 @@ func (bot *CQBot) CQProcessGroupRequest(flag, subType, reason string, approve bo
 			}
 		}
 	}
-	log.Errorf("处理群系统消息失败: 消息 %v 不存在.", flag)
+	log.Warnf("处理群系统消息失败: 消息 %v 不存在.", flag)
 	return Failed(100, "FLAG_NOT_FOUND", "FLAG不存在")
 }
 
 // CQDeleteMessage 撤回消息
 //
 // https:// git.io/Jtz1y
-func (bot *CQBot) CQDeleteMessage(messageID int32) MSG {
-	msg := bot.GetMessage(messageID)
-	if msg == nil {
+// @route(delete_msg)
+func (bot *CQBot) CQDeleteMessage(messageID int32) global.MSG {
+	msg, err := db.GetMessageByGlobalID(messageID)
+	if err != nil {
+		log.Warnf("撤回消息时出现错误: %v", err)
 		return Failed(100, "MESSAGE_NOT_FOUND", "消息不存在")
 	}
-	if _, ok := msg["group"]; ok {
-		if msg["internal-id"] == nil {
-			// TODO 撤回临时对话消息
-			log.Warnf("撤回 %v 失败: 无法撤回临时对话消息", messageID)
-			return Failed(100, "CANNOT_RECALL_TEMP_MSG", "无法撤回临时对话消息")
-		}
-		if err := bot.Client.RecallGroupMessage(msg["group"].(int64), msg["message-id"].(int32), msg["internal-id"].(int32)); err != nil {
+	switch o := msg.(type) {
+	case *db.StoredGroupMessage:
+		if err = bot.Client.RecallGroupMessage(o.GroupCode, o.Attribute.MessageSeq, o.Attribute.InternalID); err != nil {
 			log.Warnf("撤回 %v 失败: %v", messageID, err)
 			return Failed(100, "RECALL_API_ERROR", err.Error())
 		}
-	} else {
-		if msg["sender"].(message.Sender).Uin != bot.Client.Uin {
+	case *db.StoredPrivateMessage:
+		if o.Attribute.SenderUin != bot.Client.Uin {
 			log.Warnf("撤回 %v 失败: 好友会话无法撤回对方消息.", messageID)
 			return Failed(100, "CANNOT_RECALL_FRIEND_MSG", "无法撤回对方消息")
 		}
-		if err := bot.Client.RecallPrivateMessage(msg["target"].(int64), int64(msg["time"].(int32)), msg["message-id"].(int32), msg["internal-id"].(int32)); err != nil {
+		if err = bot.Client.RecallPrivateMessage(o.TargetUin, o.Attribute.Timestamp, o.Attribute.MessageSeq, o.Attribute.InternalID); err != nil {
 			log.Warnf("撤回 %v 失败: %v", messageID, err)
 			return Failed(100, "RECALL_API_ERROR", err.Error())
 		}
+	default:
+		return Failed(100, "UNKNOWN_ERROR")
 	}
 	return OK(nil)
 }
@@ -722,7 +1202,9 @@ func (bot *CQBot) CQDeleteMessage(messageID int32) MSG {
 // CQSetGroupAdmin 群组设置管理员
 //
 // https://git.io/Jtz1S
-func (bot *CQBot) CQSetGroupAdmin(groupID, userID int64, enable bool) MSG {
+// @route(set_group_admin)
+// @default(enable=true)
+func (bot *CQBot) CQSetGroupAdmin(groupID, userID int64, enable bool) global.MSG {
 	group := bot.Client.FindGroup(groupID)
 	if group == nil || group.OwnerUin != bot.Client.Uin {
 		return Failed(100, "PERMISSION_DENIED", "群不存在或权限不足")
@@ -744,12 +1226,13 @@ func (bot *CQBot) CQSetGroupAdmin(groupID, userID int64, enable bool) MSG {
 // CQGetVipInfo 扩展API-获取VIP信息
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96vip%E4%BF%A1%E6%81%AF
-func (bot *CQBot) CQGetVipInfo(userID int64) MSG {
+// @route(_get_vip_info)
+func (bot *CQBot) CQGetVipInfo(userID int64) global.MSG {
 	vip, err := bot.Client.GetVipInfo(userID)
 	if err != nil {
 		return Failed(100, "VIP_API_ERROR", err.Error())
 	}
-	msg := MSG{
+	msg := global.MSG{
 		"user_id":          vip.Uin,
 		"nickname":         vip.Name,
 		"level":            vip.Level,
@@ -764,11 +1247,13 @@ func (bot *CQBot) CQGetVipInfo(userID int64) MSG {
 // CQGetGroupHonorInfo 获取群荣誉信息
 //
 // https://git.io/Jtz1H
-func (bot *CQBot) CQGetGroupHonorInfo(groupID int64, t string) MSG {
-	msg := MSG{"group_id": groupID}
-	convertMem := func(memList []client.HonorMemberInfo) (ret []MSG) {
+// @route(get_group_honor_info)
+// @rename(t->type)
+func (bot *CQBot) CQGetGroupHonorInfo(groupID int64, t string) global.MSG {
+	msg := global.MSG{"group_id": groupID}
+	convertMem := func(memList []client.HonorMemberInfo) (ret []global.MSG) {
 		for _, mem := range memList {
-			ret = append(ret, MSG{
+			ret = append(ret, global.MSG{
 				"user_id":     mem.Uin,
 				"nickname":    mem.Name,
 				"avatar":      mem.Avatar,
@@ -780,7 +1265,7 @@ func (bot *CQBot) CQGetGroupHonorInfo(groupID int64, t string) MSG {
 	if t == "talkative" || t == "all" {
 		if honor, err := bot.Client.GetGroupHonorInfo(groupID, client.Talkative); err == nil {
 			if honor.CurrentTalkative.Uin != 0 {
-				msg["current_talkative"] = MSG{
+				msg["current_talkative"] = global.MSG{
 					"user_id":   honor.CurrentTalkative.Uin,
 					"nickname":  honor.CurrentTalkative.Name,
 					"avatar":    honor.CurrentTalkative.Avatar,
@@ -821,12 +1306,13 @@ func (bot *CQBot) CQGetGroupHonorInfo(groupID int64, t string) MSG {
 // CQGetStrangerInfo 获取陌生人信息
 //
 // https://git.io/Jtz17
-func (bot *CQBot) CQGetStrangerInfo(userID int64) MSG {
+// @route(get_stranger_info)
+func (bot *CQBot) CQGetStrangerInfo(userID int64) global.MSG {
 	info, err := bot.Client.GetSummaryInfo(userID)
 	if err != nil {
 		return Failed(100, "SUMMARY_API_ERROR", err.Error())
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"user_id":  info.Uin,
 		"nickname": info.Nickname,
 		"qid":      info.Qid,
@@ -848,7 +1334,8 @@ func (bot *CQBot) CQGetStrangerInfo(userID int64) MSG {
 // CQHandleQuickOperation 隐藏API-对事件执行快速操作
 //
 // https://git.io/Jtz15
-func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) MSG {
+// @route(".handle_quick_operation")
+func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) global.MSG {
 	postType := context.Get("post_type").Str
 
 	switch postType {
@@ -859,20 +1346,20 @@ func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) MSG {
 		reply := operation.Get("reply")
 
 		if reply.Exists() {
-			autoEscape := global.EnsureBool(operation.Get("auto_escape"), false)
+			autoEscape := param.EnsureBool(operation.Get("auto_escape"), false)
 			at := operation.Get("at_sender").Bool() && !isAnonymous && msgType == "group"
 			if at && reply.IsArray() {
 				// 在 reply 数组头部插入CQ码
-				replySegments := make([]MSG, 0)
-				segments := make([]MSG, 0)
-				segments = append(segments, MSG{
+				replySegments := make([]global.MSG, 0)
+				segments := make([]global.MSG, 0)
+				segments = append(segments, global.MSG{
 					"type": "at",
-					"data": MSG{
+					"data": global.MSG{
 						"qq": context.Get("sender.user_id").Int(),
 					},
 				})
 
-				err := json.UnmarshalFromString(reply.Raw, &replySegments)
+				err := json.Unmarshal(utils.S2B(reply.Raw), &replySegments)
 				if err != nil {
 					log.WithError(err).Warnf("处理 at_sender 过程中发生错误")
 					return Failed(-1, "处理 at_sender 过程中发生错误", err.Error())
@@ -880,13 +1367,13 @@ func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) MSG {
 
 				segments = append(segments, replySegments...)
 
-				modified, err := json.MarshalToString(segments)
+				modified, err := json.Marshal(segments)
 				if err != nil {
 					log.WithError(err).Warnf("处理 at_sender 过程中发生错误")
 					return Failed(-1, "处理 at_sender 过程中发生错误", err.Error())
 				}
 
-				reply = gjson.Parse(modified)
+				reply = gjson.Parse(utils.B2S(modified))
 			} else if at && reply.Type == gjson.String {
 				reply = gjson.Parse(fmt.Sprintf(
 					"\"[CQ:at,qq=%d]%s\"",
@@ -924,10 +1411,10 @@ func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) MSG {
 		reqType := context.Get("request_type").Str
 		if operation.Get("approve").Exists() {
 			if reqType == "friend" {
-				bot.CQProcessFriendRequest(context.Get("flag").Str, operation.Get("approve").Bool())
+				bot.CQProcessFriendRequest(context.Get("flag").String(), operation.Get("approve").Bool())
 			}
 			if reqType == "group" {
-				bot.CQProcessGroupRequest(context.Get("flag").Str, context.Get("sub_type").Str, operation.Get("reason").Str, operation.Get("approve").Bool())
+				bot.CQProcessGroupRequest(context.Get("flag").String(), context.Get("sub_type").Str, operation.Get("reason").Str, operation.Get("approve").Bool())
 			}
 		}
 	}
@@ -937,23 +1424,41 @@ func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) MSG {
 // CQGetImage 获取图片(修改自OneBot)
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E5%9B%BE%E7%89%87%E4%BF%A1%E6%81%AF
-func (bot *CQBot) CQGetImage(file string) MSG {
-	if !global.PathExists(path.Join(global.ImagePath, file)) {
-		return Failed(100)
+// @route(get_image)
+func (bot *CQBot) CQGetImage(file string) global.MSG {
+	var b []byte
+	var err error
+	if strings.HasSuffix(file, ".image") {
+		var f []byte
+		f, err = hex.DecodeString(strings.TrimSuffix(file, ".image"))
+		b = cache.Image.Get(f)
 	}
-	b, err := ioutil.ReadFile(path.Join(global.ImagePath, file))
+
+	if b == nil {
+		if !global.PathExists(path.Join(global.ImagePath, file)) {
+			return Failed(100)
+		}
+		b, err = os.ReadFile(path.Join(global.ImagePath, file))
+	}
+
 	if err == nil {
 		r := binary.NewReader(b)
 		r.ReadBytes(16)
-		msg := MSG{
+		msg := global.MSG{
 			"size":     r.ReadInt32(),
 			"filename": r.ReadString(),
 			"url":      r.ReadString(),
 		}
-		local := path.Join(global.CachePath, file+"."+path.Ext(msg["filename"].(string)))
+		local := path.Join(global.CachePath, file+path.Ext(msg["filename"].(string)))
 		if !global.PathExists(local) {
-			if data, err := global.GetBytes(msg["url"].(string)); err == nil {
-				_ = ioutil.WriteFile(local, data, 0644)
+			if body, err := global.HTTPGetReadCloser(msg["url"].(string)); err == nil {
+				f, _ := os.OpenFile(local, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0o0644)
+				_, _ = f.ReadFrom(body)
+				_ = body.Close()
+				_ = f.Close()
+			} else {
+				log.Warnf("下载图片 %v 时出现错误: %v", msg["url"], err)
+				return Failed(100, "DOWNLOAD_IMAGE_ERROR", err.Error())
 			}
 		}
 		msg["file"] = local
@@ -965,7 +1470,27 @@ func (bot *CQBot) CQGetImage(file string) MSG {
 // CQDownloadFile 扩展API-下载文件到缓存目录
 //
 // https://docs.go-cqhttp.org/api/#%E4%B8%8B%E8%BD%BD%E6%96%87%E4%BB%B6%E5%88%B0%E7%BC%93%E5%AD%98%E7%9B%AE%E5%BD%95
-func (bot *CQBot) CQDownloadFile(url string, headers map[string]string, threadCount int) MSG {
+// @route(download_file)
+func (bot *CQBot) CQDownloadFile(url string, headers gjson.Result, threadCount int) global.MSG {
+	h := map[string]string{}
+	if headers.IsArray() {
+		for _, sub := range headers.Array() {
+			str := strings.SplitN(sub.String(), "=", 2)
+			if len(str) == 2 {
+				h[str[0]] = str[1]
+			}
+		}
+	}
+	if headers.Type == gjson.String {
+		lines := strings.Split(headers.String(), "\r\n")
+		for _, sub := range lines {
+			str := strings.SplitN(sub, "=", 2)
+			if len(str) == 2 {
+				h[str[0]] = str[1]
+			}
+		}
+	}
+
 	hash := md5.Sum([]byte(url))
 	file := path.Join(global.CachePath, hex.EncodeToString(hash[:])+".cache")
 	if global.PathExists(file) {
@@ -974,12 +1499,12 @@ func (bot *CQBot) CQDownloadFile(url string, headers map[string]string, threadCo
 			return Failed(100, "DELETE_FILE_ERROR", err.Error())
 		}
 	}
-	if err := global.DownloadFileMultiThreading(url, file, 0, threadCount, headers); err != nil {
+	if err := global.DownloadFileMultiThreading(url, file, 0, threadCount, h); err != nil {
 		log.Warnf("下载链接 %v 时出现错误: %v", url, err)
 		return Failed(100, "DOWNLOAD_FILE_ERROR", err.Error())
 	}
 	abs, _ := filepath.Abs(file)
-	return OK(MSG{
+	return OK(global.MSG{
 		"file": abs,
 	})
 }
@@ -987,70 +1512,143 @@ func (bot *CQBot) CQDownloadFile(url string, headers map[string]string, threadCo
 // CQGetForwardMessage 获取合并转发消息
 //
 // https://git.io/Jtz1F
-func (bot *CQBot) CQGetForwardMessage(resID string) MSG {
+// @route(get_forward_msg)
+// @rename(res_id->"[message_id\x2Cid].0")
+func (bot *CQBot) CQGetForwardMessage(resID string) global.MSG {
 	m := bot.Client.GetForwardMessage(resID)
 	if m == nil {
 		return Failed(100, "MSG_NOT_FOUND", "消息不存在")
 	}
-	r := make([]MSG, 0, len(m.Nodes))
-	for _, n := range m.Nodes {
-		bot.checkMedia(n.Message)
-		r = append(r, MSG{
-			"sender": MSG{
-				"user_id":  n.SenderId,
-				"nickname": n.SenderName,
-			},
-			"time":    n.Time,
-			"content": ToFormattedMessage(n.Message, 0, false),
-		})
+
+	var transformNodes func(nodes []*message.ForwardNode) []global.MSG
+	transformNodes = func(nodes []*message.ForwardNode) []global.MSG {
+		r := make([]global.MSG, len(nodes))
+		for i, n := range nodes {
+			bot.checkMedia(n.Message, 0)
+			content := ToFormattedMessage(n.Message, MessageSource{SourceType: MessageSourceGroup}, false)
+			if len(n.Message) == 1 {
+				if forward, ok := n.Message[0].(*message.ForwardMessage); ok {
+					content = transformNodes(forward.Nodes)
+				}
+			}
+			r[i] = global.MSG{
+				"sender": global.MSG{
+					"user_id":  n.SenderId,
+					"nickname": n.SenderName,
+				},
+				"time":    n.Time,
+				"content": content,
+			}
+		}
+		return r
 	}
-	return OK(MSG{
-		"messages": r,
+	return OK(global.MSG{
+		"messages": transformNodes(m.Nodes),
 	})
 }
 
 // CQGetMessage 获取消息
 //
 // https://git.io/Jtz1b
-func (bot *CQBot) CQGetMessage(messageID int32) MSG {
-	msg := bot.GetMessage(messageID)
-	if msg == nil {
+// @route(get_msg)
+func (bot *CQBot) CQGetMessage(messageID int32) global.MSG {
+	msg, err := db.GetMessageByGlobalID(messageID)
+	if err != nil {
+		log.Warnf("获取消息时出现错误: %v", err)
 		return Failed(100, "MSG_NOT_FOUND", "消息不存在")
 	}
-	sender := msg["sender"].(message.Sender)
-	gid, isGroup := msg["group"]
-	raw := msg["message"].(string)
-	return OK(MSG{
-		"message_id":  messageID,
-		"real_id":     msg["message-id"],
-		"message_seq": msg["message-id"],
-		"group":       isGroup,
-		"group_id":    gid,
-		"message_type": func() string {
-			if isGroup {
-				return "group"
-			}
-			return "private"
-		}(),
-		"sender": MSG{
-			"user_id":  sender.Uin,
-			"nickname": sender.Nickname,
+	m := global.MSG{
+		"message_id":    msg.GetGlobalID(),
+		"message_id_v2": msg.GetID(),
+		"message_type":  msg.GetType(),
+		"real_id":       msg.GetAttribute().MessageSeq,
+		"message_seq":   msg.GetAttribute().MessageSeq,
+		"group":         msg.GetType() == "group",
+		"sender": global.MSG{
+			"user_id":  msg.GetAttribute().SenderUin,
+			"nickname": msg.GetAttribute().SenderName,
 		},
-		"time":        msg["time"],
-		"raw_message": raw,
-		"message": ToFormattedMessage(bot.ConvertStringMessage(raw, isGroup), func() int64 {
-			if isGroup {
-				return gid.(int64)
+		"time": msg.GetAttribute().Timestamp,
+	}
+	switch o := msg.(type) {
+	case *db.StoredGroupMessage:
+		m["group_id"] = o.GroupCode
+		m["message"] = ToFormattedMessage(bot.ConvertContentMessage(o.Content, MessageSourceGroup), MessageSource{SourceType: MessageSourceGroup, PrimaryID: uint64(o.GroupCode)}, false)
+	case *db.StoredPrivateMessage:
+		m["message"] = ToFormattedMessage(bot.ConvertContentMessage(o.Content, MessageSourcePrivate), MessageSource{SourceType: MessageSourcePrivate}, false)
+	}
+	return OK(m)
+}
+
+// CQGetGuildMessage 获取频道消息
+// @route(get_guild_msg)
+func (bot *CQBot) CQGetGuildMessage(messageID string, noCache bool) global.MSG {
+	source, seq := decodeGuildMessageID(messageID)
+	if source == nil {
+		log.Warnf("获取消息时出现错误: 无效消息ID")
+		return Failed(100, "INVALID_MESSAGE_ID", "无效消息ID")
+	}
+	m := global.MSG{
+		"message_id": messageID,
+		"message_source": func() string {
+			if source.SourceType == MessageSourceGuildDirect {
+				return "direct"
 			}
-			return sender.Uin
-		}(), false),
-	})
+			return "channel"
+		}(),
+		"message_seq": seq,
+		"guild_id":    fU64(source.PrimaryID),
+		"reactions":   []int{},
+	}
+	// nolint: exhaustive
+	switch source.SourceType {
+	case MessageSourceGuildChannel:
+		m["channel_id"] = fU64(source.SubID)
+		if noCache {
+			pull, err := bot.Client.GuildService.PullGuildChannelMessage(source.PrimaryID, source.SubID, seq, seq)
+			if err != nil {
+				log.Warnf("获取消息时出现错误: %v", err)
+				return Failed(100, "API_ERROR", err.Error())
+			}
+			if len(m) == 0 {
+				log.Warnf("获取消息时出现错误: 消息不存在")
+				return Failed(100, "MSG_NOT_FOUND", "消息不存在")
+			}
+			m["time"] = pull[0].Time
+			m["sender"] = global.MSG{
+				"user_id":  pull[0].Sender.TinyId,
+				"tiny_id":  fU64(pull[0].Sender.TinyId),
+				"nickname": pull[0].Sender.Nickname,
+			}
+			m["message"] = ToFormattedMessage(pull[0].Elements, *source, false)
+			m["reactions"] = convertReactions(pull[0].Reactions)
+			bot.InsertGuildChannelMessage(pull[0])
+		} else {
+			channelMsgByDB, err := db.GetGuildChannelMessageByID(messageID)
+			if err != nil {
+				log.Warnf("获取消息时出现错误: %v", err)
+				return Failed(100, "MSG_NOT_FOUND", "消息不存在")
+			}
+			m["time"] = channelMsgByDB.Attribute.Timestamp
+			m["sender"] = global.MSG{
+				"user_id":  channelMsgByDB.Attribute.SenderTinyID,
+				"tiny_id":  fU64(channelMsgByDB.Attribute.SenderTinyID),
+				"nickname": channelMsgByDB.Attribute.SenderName,
+			}
+			m["message"] = ToFormattedMessage(bot.ConvertContentMessage(channelMsgByDB.Content, MessageSourceGuildChannel), *source)
+		}
+	case MessageSourceGuildDirect:
+		// todo(mrs4s): 支持 direct 消息
+		m["tiny_id"] = fU64(source.SubID)
+	}
+	return OK(m)
 }
 
 // CQGetGroupSystemMessages 扩展API-获取群文件系统消息
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4%E7%B3%BB%E7%BB%9F%E6%B6%88%E6%81%AF
-func (bot *CQBot) CQGetGroupSystemMessages() MSG {
+// @route(get_group_system_msg)
+func (bot *CQBot) CQGetGroupSystemMessages() global.MSG {
 	msg, err := bot.Client.GetGroupSystemMessages()
 	if err != nil {
 		log.Warnf("获取群系统消息失败: %v", err)
@@ -1062,7 +1660,9 @@ func (bot *CQBot) CQGetGroupSystemMessages() MSG {
 // CQGetGroupMessageHistory 获取群消息历史记录
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4%E6%B6%88%E6%81%AF%E5%8E%86%E5%8F%B2%E8%AE%B0%E5%BD%95
-func (bot *CQBot) CQGetGroupMessageHistory(groupID int64, seq int64) MSG {
+// @route(get_group_msg_history)
+// @rename(seq->message_seq)
+func (bot *CQBot) CQGetGroupMessageHistory(groupID int64, seq int64) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g == nil {
 		return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
 	}
@@ -1078,18 +1678,15 @@ func (bot *CQBot) CQGetGroupMessageHistory(groupID int64, seq int64) MSG {
 		log.Warnf("获取群历史消息失败: %v", err)
 		return Failed(100, "MESSAGES_API_ERROR", err.Error())
 	}
-	ms := make([]MSG, 0, len(msg))
+	ms := make([]global.MSG, 0, len(msg))
 	for _, m := range msg {
-		id := m.Id
-		bot.checkMedia(m.Elements)
-		if bot.db != nil {
-			id = bot.InsertGroupMessage(m)
-		}
+		bot.checkMedia(m.Elements, groupID)
+		id := bot.InsertGroupMessage(m)
 		t := bot.formatGroupMessage(m)
 		t["message_id"] = id
 		ms = append(ms, t)
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"messages": ms,
 	})
 }
@@ -1097,22 +1694,23 @@ func (bot *CQBot) CQGetGroupMessageHistory(groupID int64, seq int64) MSG {
 // CQGetOnlineClients 扩展API-获取当前账号在线客户端列表
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E5%BD%93%E5%89%8D%E8%B4%A6%E5%8F%B7%E5%9C%A8%E7%BA%BF%E5%AE%A2%E6%88%B7%E7%AB%AF%E5%88%97%E8%A1%A8
-func (bot *CQBot) CQGetOnlineClients(noCache bool) MSG {
+// @route(get_online_clients)
+func (bot *CQBot) CQGetOnlineClients(noCache bool) global.MSG {
 	if noCache {
 		if err := bot.Client.RefreshStatus(); err != nil {
 			log.Warnf("刷新客户端状态时出现问题 %v", err)
 			return Failed(100, "REFRESH_STATUS_ERROR", err.Error())
 		}
 	}
-	d := make([]MSG, 0, len(bot.Client.OnlineClients))
+	d := make([]global.MSG, 0, len(bot.Client.OnlineClients))
 	for _, oc := range bot.Client.OnlineClients {
-		d = append(d, MSG{
+		d = append(d, global.MSG{
 			"app_id":      oc.AppId,
 			"device_name": oc.DeviceName,
 			"device_kind": oc.DeviceKind,
 		})
 	}
-	return OK(MSG{
+	return OK(global.MSG{
 		"clients": d,
 	})
 }
@@ -1120,22 +1718,26 @@ func (bot *CQBot) CQGetOnlineClients(noCache bool) MSG {
 // CQCanSendImage 检查是否可以发送图片(此处永远返回true)
 //
 // https://git.io/Jtz1N
-func (bot *CQBot) CQCanSendImage() MSG {
-	return OK(MSG{"yes": true})
+// @route(can_send_image)
+func (bot *CQBot) CQCanSendImage() global.MSG {
+	return OK(global.MSG{"yes": true})
 }
 
 // CQCanSendRecord 检查是否可以发送语音(此处永远返回true)
 //
 // https://git.io/Jtz1x
-func (bot *CQBot) CQCanSendRecord() MSG {
-	return OK(MSG{"yes": true})
+// @route(can_send_record)
+func (bot *CQBot) CQCanSendRecord() global.MSG {
+	return OK(global.MSG{"yes": true})
 }
 
 // CQOcrImage 扩展API-图片OCR
 //
 // https://docs.go-cqhttp.org/api/#%E5%9B%BE%E7%89%87-ocr
-func (bot *CQBot) CQOcrImage(imageID string) MSG {
-	img, err := bot.makeImageOrVideoElem(map[string]string{"file": imageID}, false, true)
+// @route(ocr_image,".ocr_image")
+// @rename(image_id->image)
+func (bot *CQBot) CQOcrImage(imageID string) global.MSG {
+	img, err := bot.makeImageOrVideoElem(map[string]string{"file": imageID}, false, MessageSourceGroup)
 	if err != nil {
 		log.Warnf("load image error: %v", err)
 		return Failed(100, "LOAD_FILE_ERROR", err.Error())
@@ -1151,7 +1753,8 @@ func (bot *CQBot) CQOcrImage(imageID string) MSG {
 // CQSetGroupPortrait 扩展API-设置群头像
 //
 // https://docs.go-cqhttp.org/api/#%E8%AE%BE%E7%BD%AE%E7%BE%A4%E5%A4%B4%E5%83%8F
-func (bot *CQBot) CQSetGroupPortrait(groupID int64, file, cache string) MSG {
+// @route(set_group_portrait)
+func (bot *CQBot) CQSetGroupPortrait(groupID int64, file, cache string) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
 		img, err := global.FindFile(file, cache, global.ImagePath)
 		if err != nil {
@@ -1167,7 +1770,9 @@ func (bot *CQBot) CQSetGroupPortrait(groupID int64, file, cache string) MSG {
 // CQSetGroupAnonymousBan 群组匿名用户禁言
 //
 // https://git.io/Jtz1p
-func (bot *CQBot) CQSetGroupAnonymousBan(groupID int64, flag string, duration int32) MSG {
+// @route(set_group_anonymous_ban)
+// @rename(flag->"[anonymous_flag\x2Canonymous.flag].0")
+func (bot *CQBot) CQSetGroupAnonymousBan(groupID int64, flag string, duration int32) global.MSG {
 	if flag == "" {
 		return Failed(100, "INVALID_FLAG", "无效的flag")
 	}
@@ -1190,14 +1795,15 @@ func (bot *CQBot) CQSetGroupAnonymousBan(groupID int64, flag string, duration in
 // CQGetStatus 获取运行状态
 //
 // https://git.io/JtzMe
-func (bot *CQBot) CQGetStatus() MSG {
-	return OK(MSG{
+// @route(get_status)
+func (bot *CQBot) CQGetStatus() global.MSG {
+	return OK(global.MSG{
 		"app_initialized": true,
 		"app_enabled":     true,
 		"plugins_good":    nil,
 		"app_good":        true,
-		"online":          bot.Client.Online,
-		"good":            bot.Client.Online,
+		"online":          bot.Client.Online.Load(),
+		"good":            bot.Client.Online.Load(),
 		"stat":            bot.Client.GetStatistics(),
 	})
 }
@@ -1205,19 +1811,15 @@ func (bot *CQBot) CQGetStatus() MSG {
 // CQSetEssenceMessage 扩展API-设置精华消息
 //
 // https://docs.go-cqhttp.org/api/#%E8%AE%BE%E7%BD%AE%E7%B2%BE%E5%8D%8E%E6%B6%88%E6%81%AF
-func (bot *CQBot) CQSetEssenceMessage(messageID int32) MSG {
-	msg := bot.GetMessage(messageID)
-	if msg == nil {
+// @route(set_essence_msg)
+func (bot *CQBot) CQSetEssenceMessage(messageID int32) global.MSG {
+	msg, err := db.GetGroupMessageByGlobalID(messageID)
+	if err != nil {
 		return Failed(100, "MESSAGE_NOT_FOUND", "消息不存在")
 	}
-	if _, ok := msg["group"]; ok {
-		if err := bot.Client.SetEssenceMessage(msg["group"].(int64), msg["message-id"].(int32), msg["internal-id"].(int32)); err != nil {
-			log.Warnf("设置精华消息 %v 失败: %v", messageID, err)
-			return Failed(100, "SET_ESSENCE_MSG_ERROR", err.Error())
-		}
-	} else {
-		log.Warnf("设置精华消息 %v 失败: 非群聊", messageID)
-		return Failed(100, "SET_ESSENCE_MSG_ERROR", "非群聊")
+	if err := bot.Client.SetEssenceMessage(msg.GroupCode, msg.Attribute.MessageSeq, msg.Attribute.InternalID); err != nil {
+		log.Warnf("设置精华消息 %v 失败: %v", messageID, err)
+		return Failed(100, "SET_ESSENCE_MSG_ERROR", err.Error())
 	}
 	return OK(nil)
 }
@@ -1225,19 +1827,15 @@ func (bot *CQBot) CQSetEssenceMessage(messageID int32) MSG {
 // CQDeleteEssenceMessage 扩展API-移出精华消息
 //
 // https://docs.go-cqhttp.org/api/#%E7%A7%BB%E5%87%BA%E7%B2%BE%E5%8D%8E%E6%B6%88%E6%81%AF
-func (bot *CQBot) CQDeleteEssenceMessage(messageID int32) MSG {
-	msg := bot.GetMessage(messageID)
-	if msg == nil {
+// @route(delete_essence_msg)
+func (bot *CQBot) CQDeleteEssenceMessage(messageID int32) global.MSG {
+	msg, err := db.GetGroupMessageByGlobalID(messageID)
+	if err != nil {
 		return Failed(100, "MESSAGE_NOT_FOUND", "消息不存在")
 	}
-	if _, ok := msg["group"]; ok {
-		if err := bot.Client.DeleteEssenceMessage(msg["group"].(int64), msg["message-id"].(int32), msg["internal-id"].(int32)); err != nil {
-			log.Warnf("移出精华消息 %v 失败: %v", messageID, err)
-			return Failed(100, "DEL_ESSENCE_MSG_ERROR", err.Error())
-		}
-	} else {
-		log.Warnf("移出精华消息 %v 失败: 非群聊", messageID)
-		return Failed(100, "DEL_ESSENCE_MSG_ERROR", "非群聊")
+	if err := bot.Client.DeleteEssenceMessage(msg.GroupCode, msg.Attribute.MessageSeq, msg.Attribute.InternalID); err != nil {
+		log.Warnf("删除精华消息 %v 失败: %v", messageID, err)
+		return Failed(100, "SET_ESSENCE_MSG_ERROR", err.Error())
 	}
 	return OK(nil)
 }
@@ -1245,26 +1843,27 @@ func (bot *CQBot) CQDeleteEssenceMessage(messageID int32) MSG {
 // CQGetEssenceMessageList 扩展API-获取精华消息列表
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%B2%BE%E5%8D%8E%E6%B6%88%E6%81%AF%E5%88%97%E8%A1%A8
-func (bot *CQBot) CQGetEssenceMessageList(groupCode int64) MSG {
-	g := bot.Client.FindGroup(groupCode)
+// @route(get_essence_msg_list)
+func (bot *CQBot) CQGetEssenceMessageList(groupID int64) global.MSG {
+	g := bot.Client.FindGroup(groupID)
 	if g == nil {
 		return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
 	}
-	msgList, err := bot.Client.GetGroupEssenceMsgList(groupCode)
+	msgList, err := bot.Client.GetGroupEssenceMsgList(groupID)
 	if err != nil {
 		return Failed(100, "GET_ESSENCE_LIST_FOUND", err.Error())
 	}
-	list := make([]MSG, 0, len(msgList))
+	list := make([]global.MSG, 0, len(msgList))
 	for _, m := range msgList {
-		msg := MSG{
+		msg := global.MSG{
 			"sender_nick":   m.SenderNick,
 			"sender_time":   m.SenderTime,
 			"operator_time": m.AddDigestTime,
 			"operator_nick": m.AddDigestNick,
+			"sender_id":     m.SenderUin,
+			"operator_id":   m.AddDigestUin,
 		}
-		msg["sender_id"], _ = strconv.ParseUint(m.SenderUin, 10, 64)
-		msg["operator_id"], _ = strconv.ParseUint(m.AddDigestUin, 10, 64)
-		msg["message_id"] = toGlobalID(groupCode, int32(m.MessageID))
+		msg["message_id"] = db.ToGlobalID(groupID, int32(m.MessageID))
 		list = append(list, msg)
 	}
 	return OK(list)
@@ -1273,8 +1872,9 @@ func (bot *CQBot) CQGetEssenceMessageList(groupCode int64) MSG {
 // CQCheckURLSafely 扩展API-检查链接安全性
 //
 // https://docs.go-cqhttp.org/api/#%E6%A3%80%E6%9F%A5%E9%93%BE%E6%8E%A5%E5%AE%89%E5%85%A8%E6%80%A7
-func (bot *CQBot) CQCheckURLSafely(url string) MSG {
-	return OK(MSG{
+// @route(check_url_safely)
+func (bot *CQBot) CQCheckURLSafely(url string) global.MSG {
+	return OK(global.MSG{
 		"level": bot.Client.CheckUrlSafely(url),
 	})
 }
@@ -1282,12 +1882,13 @@ func (bot *CQBot) CQCheckURLSafely(url string) MSG {
 // CQGetVersionInfo 获取版本信息
 //
 // https://git.io/JtwUs
-func (bot *CQBot) CQGetVersionInfo() MSG {
+// @route(get_version_info)
+func (bot *CQBot) CQGetVersionInfo() global.MSG {
 	wd, _ := os.Getwd()
-	return OK(MSG{
+	return OK(global.MSG{
 		"app_name":                   "go-cqhttp",
-		"app_version":                Version,
-		"app_full_name":              fmt.Sprintf("go-cqhttp-%s_%s_%s-%s", Version, runtime.GOOS, runtime.GOARCH, runtime.Version()),
+		"app_version":                base.Version,
+		"app_full_name":              fmt.Sprintf("go-cqhttp-%s_%s_%s-%s", base.Version, runtime.GOOS, runtime.GOARCH, runtime.Version()),
 		"protocol_version":           "v11",
 		"coolq_directory":            wd,
 		"coolq_edition":              "pro",
@@ -1297,10 +1898,10 @@ func (bot *CQBot) CQGetVersionInfo() MSG {
 		"plugin_build_configuration": "release",
 		"runtime_version":            runtime.Version(),
 		"runtime_os":                 runtime.GOOS,
-		"version":                    Version,
+		"version":                    base.Version,
 		"protocol": func() int {
 			switch client.SystemDeviceInfo.Protocol {
-			case client.IPad:
+			case client.Unset, client.IPad:
 				return 0
 			case client.AndroidPhone:
 				return 1
@@ -1308,6 +1909,8 @@ func (bot *CQBot) CQGetVersionInfo() MSG {
 				return 2
 			case client.MacOS:
 				return 3
+			case client.QiDian:
+				return 4
 			default:
 				return -1
 			}
@@ -1315,13 +1918,72 @@ func (bot *CQBot) CQGetVersionInfo() MSG {
 	})
 }
 
+// CQGetModelShow 获取在线机型
+//
+// https://club.vip.qq.com/onlinestatus/set
+// @route(_get_model_show)
+func (bot *CQBot) CQGetModelShow(model string) global.MSG {
+	variants, err := bot.Client.GetModelShow(model)
+	if err != nil {
+		return Failed(100, "GET_MODEL_SHOW_API_ERROR", "无法获取在线机型")
+	}
+	a := make([]global.MSG, 0, len(variants))
+	for _, v := range variants {
+		a = append(a, global.MSG{
+			"model_show": v.ModelShow,
+			"need_pay":   v.NeedPay,
+		})
+	}
+	return OK(global.MSG{
+		"variants": a,
+	})
+}
+
+// CQSetModelShow 设置在线机型
+//
+// https://club.vip.qq.com/onlinestatus/set
+// @route(_set_model_show)
+func (bot *CQBot) CQSetModelShow(model, modelShow string) global.MSG {
+	err := bot.Client.SetModelShow(model, modelShow)
+	if err != nil {
+		return Failed(100, "SET_MODEL_SHOW_API_ERROR", "无法设置在线机型")
+	}
+	return OK(nil)
+}
+
+// CQMarkMessageAsRead 标记消息已读
+// @route(mark_msg_as_read)
+// @rename(msg_id->message_id)
+func (bot *CQBot) CQMarkMessageAsRead(msgID int32) global.MSG {
+	m, err := db.GetMessageByGlobalID(msgID)
+	if err != nil {
+		return Failed(100, "MSG_NOT_FOUND", "消息不存在")
+	}
+	switch o := m.(type) {
+	case *db.StoredGroupMessage:
+		bot.Client.MarkGroupMessageReaded(o.GroupCode, int64(o.Attribute.MessageSeq))
+		return OK(nil)
+	case *db.StoredPrivateMessage:
+		bot.Client.MarkPrivateMessageReaded(o.SessionUin, o.Attribute.Timestamp)
+	}
+	return OK(nil)
+}
+
+// CQReloadEventFilter 重载事件过滤器
+//
+// @route(reload_event_filter)
+func (bot *CQBot) CQReloadEventFilter(file string) global.MSG {
+	filter.Add(file)
+	return OK(nil)
+}
+
 // OK 生成成功返回值
-func OK(data interface{}) MSG {
-	return MSG{"data": data, "retcode": 0, "status": "ok"}
+func OK(data interface{}) global.MSG {
+	return global.MSG{"data": data, "retcode": 0, "status": "ok"}
 }
 
 // Failed 生成失败返回值
-func Failed(code int, msg ...string) MSG {
+func Failed(code int, msg ...string) global.MSG {
 	m := ""
 	w := ""
 	if len(msg) > 0 {
@@ -1330,58 +1992,21 @@ func Failed(code int, msg ...string) MSG {
 	if len(msg) > 1 {
 		w = msg[1]
 	}
-	return MSG{"data": nil, "retcode": code, "msg": m, "wording": w, "status": "failed"}
-}
-
-func convertGroupMemberInfo(groupID int64, m *client.GroupMemberInfo) MSG {
-	return MSG{
-		"group_id": groupID,
-		"user_id":  m.Uin,
-		"nickname": m.Nickname,
-		"card":     m.CardName,
-		"sex": func() string {
-			if m.Gender == 1 {
-				return "female"
-			} else if m.Gender == 0 {
-				return "male"
-			}
-			// unknown = 0xff
-			return "unknown"
-		}(),
-		"age":            0,
-		"area":           "",
-		"join_time":      m.JoinTime,
-		"last_sent_time": m.LastSpeakTime,
-		"level":          strconv.FormatInt(int64(m.Level), 10),
-		"role": func() string {
-			switch m.Permission {
-			case client.Owner:
-				return "owner"
-			case client.Administrator:
-				return "admin"
-			case client.Member:
-				return "member"
-			default:
-				return "member"
-			}
-		}(),
-		"unfriendly":        false,
-		"title":             m.SpecialTitle,
-		"title_expire_time": m.SpecialTitleExpireTime,
-		"card_changeable":   false,
-	}
+	return global.MSG{"data": nil, "retcode": code, "msg": m, "wording": w, "status": "failed"}
 }
 
 func limitedString(str string) string {
-	if utf8.RuneCountInString(str) <= 10 {
+	limited := [14]rune{10: ' ', 11: '.', 12: '.', 13: '.'}
+	i := 0
+	for _, r := range str {
+		if i >= 10 {
+			break
+		}
+		limited[i] = r
+		i++
+	}
+	if i != 10 {
 		return str
 	}
-	b := utils.S2B(str)
-	limited := make([]rune, 0, 10)
-	for i := 0; i < 10; i++ {
-		decodeRune, size := utf8.DecodeRune(b)
-		b = b[size:]
-		limited = append(limited, decodeRune)
-	}
-	return string(limited) + " ..."
+	return string(limited[:])
 }
