@@ -3,23 +3,19 @@ package coolq
 import (
 	"bytes"
 	"crypto/md5"
-	"encoding/base64"
 	"encoding/hex"
 	xml2 "encoding/xml"
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"math/rand"
 	"net/url"
 	"os"
 	"path"
-	"reflect"
 	"runtime"
 	"strconv"
 	"strings"
 	"time"
-	"unsafe"
 
 	"github.com/Mrs4s/MiraiGo/binary"
 	"github.com/Mrs4s/MiraiGo/message"
@@ -27,7 +23,11 @@ import (
 	log "github.com/sirupsen/logrus"
 	"github.com/tidwall/gjson"
 
+	"github.com/Mrs4s/go-cqhttp/db"
 	"github.com/Mrs4s/go-cqhttp/global"
+	"github.com/Mrs4s/go-cqhttp/internal/base"
+	"github.com/Mrs4s/go-cqhttp/internal/cache"
+	"github.com/Mrs4s/go-cqhttp/internal/param"
 )
 
 /*
@@ -36,52 +36,18 @@ var typeReg = regexp.MustCompile(`\[CQ:(\w+)`)
 var paramReg = regexp.MustCompile(`,([\w\-.]+?)=([^,\]]+)`)
 */
 
-// RemoveReplyAt 是否删除reply后的at
-var RemoveReplyAt bool
-
-// ExtraReplyData 是否上报额外reply信息
-var ExtraReplyData bool
-
-// IgnoreInvalidCQCode 是否忽略无效CQ码
-var IgnoreInvalidCQCode = false
-
-// SplitURL 是否分割URL
-var SplitURL = false
-
-// magicCQ 代表 uint32([]byte("[CQ:"))
-var magicCQ = uint32(0)
-
-func init() {
-	CQHeader := "[CQ:"
-	magicCQ = *(*uint32)(unsafe.Pointer((*reflect.StringHeader)(unsafe.Pointer(&CQHeader)).Data))
-}
-
-// add 指针运算
-func add(ptr unsafe.Pointer, offset uintptr) unsafe.Pointer {
-	return unsafe.Pointer(uintptr(ptr) + offset)
-}
-
-const (
-	maxImageSize = 1024 * 1024 * 30  // 30MB
-	maxVideoSize = 1024 * 1024 * 100 // 100MB
-)
-
 // PokeElement 拍一拍
 type PokeElement struct {
 	Target int64
 }
 
-// GiftElement 礼物
-type GiftElement struct {
-	Target int64
-	GiftID message.GroupGift
-}
-
 // LocalImageElement 本地图片
 type LocalImageElement struct {
-	message.ImageElement
 	Stream io.ReadSeeker
 	File   string
+
+	Flash    bool
+	EffectID int32
 }
 
 // LocalVoiceElement 本地语音
@@ -92,33 +58,43 @@ type LocalVoiceElement struct {
 
 // LocalVideoElement 本地视频
 type LocalVideoElement struct {
-	message.ShortVideoElement
 	File  string
 	thumb io.ReadSeeker
 }
 
-// Type 获取元素类型ID
-func (e *GiftElement) Type() message.ElementType {
-	// Make message.IMessageElement Happy
-	return message.At
+// MessageSource 消息来源
+// 如果为私聊或者群聊, PrimaryID 将代表群号/QQ号
+// 如果为频道, PrimaryID 为 GuildID, SubID 为 ChannelID
+type MessageSource struct {
+	SourceType MessageSourceType
+	PrimaryID  uint64
+	SubID      uint64
 }
 
-// GiftID 礼物ID数组
-var GiftID = [...]message.GroupGift{
-	message.SweetWink,
-	message.HappyCola,
-	message.LuckyBracelet,
-	message.Cappuccino,
-	message.CatWatch,
-	message.FleeceGloves,
-	message.RainbowCandy,
-	message.Stronger,
-	message.LoveMicrophone,
-	message.HoldingYourHand,
-	message.CuteCat,
-	message.MysteryMask,
-	message.ImBusy,
-	message.LoveMask,
+// MessageSourceType 消息来源类型
+type MessageSourceType byte
+
+// MessageSourceType 常量
+const (
+	MessageSourcePrivate MessageSourceType = 1 << iota
+	MessageSourceGroup
+	MessageSourceGuildChannel
+	MessageSourceGuildDirect
+)
+
+const (
+	maxImageSize = 1024 * 1024 * 30  // 30MB
+	maxVideoSize = 1024 * 1024 * 100 // 100MB
+)
+
+// Type implements the message.IMessageElement.
+func (e *LocalImageElement) Type() message.ElementType {
+	return message.Image
+}
+
+// Type impl message.IMessageElement
+func (e *LocalVideoElement) Type() message.ElementType {
+	return message.Video
 }
 
 // Type 获取元素类型ID
@@ -128,169 +104,151 @@ func (e *PokeElement) Type() message.ElementType {
 }
 
 // ToArrayMessage 将消息元素数组转为MSG数组以用于消息上报
-func ToArrayMessage(e []message.IMessageElement, id int64, isRaw ...bool) (r []MSG) {
-	r = make([]MSG, 0, len(e))
-	ur := false
-	if len(isRaw) != 0 {
-		ur = isRaw[0]
-	}
+func ToArrayMessage(e []message.IMessageElement, source MessageSource) (r []global.MSG) {
+	r = make([]global.MSG, 0, len(e))
 	m := &message.SendingMessage{Elements: e}
 	reply := m.FirstOrNil(func(e message.IMessageElement) bool {
 		_, ok := e.(*message.ReplyElement)
 		return ok
 	})
-	if reply != nil {
+	if reply != nil && source.SourceType&(MessageSourceGroup|MessageSourcePrivate) != 0 {
 		replyElem := reply.(*message.ReplyElement)
-		if ExtraReplyData {
-			r = append(r, MSG{
+		rid := int64(source.PrimaryID)
+		if rid == 0 {
+			rid = replyElem.Sender
+		}
+		if replyElem.GroupID != 0 {
+			rid = replyElem.GroupID
+		}
+		if base.ExtraReplyData {
+			r = append(r, global.MSG{
 				"type": "reply",
 				"data": map[string]string{
-					"id":   fmt.Sprint(toGlobalID(id, reply.(*message.ReplyElement).ReplySeq)),
-					"seq":  string(replyElem.ReplySeq),
+					"id":   strconv.FormatInt(int64(db.ToGlobalID(rid, replyElem.ReplySeq)), 10),
+					"seq":  strconv.FormatInt(int64(replyElem.ReplySeq), 10),
 					"qq":   strconv.FormatInt(replyElem.Sender, 10),
-					"time": string(replyElem.Time),
-					"text": CQCodeEscapeValue(CQCodeEscapeText(ToStringMessage(replyElem.Elements, id))),
+					"time": strconv.FormatInt(int64(replyElem.Time), 10),
+					"text": ToStringMessage(replyElem.Elements, source),
 				},
 			})
 		} else {
-			r = append(r, MSG{
+			r = append(r, global.MSG{
 				"type": "reply",
-				"data": map[string]string{"id": fmt.Sprint(toGlobalID(id, replyElem.ReplySeq))},
+				"data": map[string]string{"id": strconv.FormatInt(int64(db.ToGlobalID(rid, replyElem.ReplySeq)), 10)},
 			})
 		}
 	}
 	for i, elem := range e {
-		var m MSG
+		var m global.MSG
 		switch o := elem.(type) {
 		case *message.ReplyElement:
-			if RemoveReplyAt && len(e) > i+1 {
+			if base.RemoveReplyAt && i+1 < len(e) {
 				elem, ok := e[i+1].(*message.AtElement)
 				if ok && elem.Target == o.Sender {
 					e[i+1] = nil
 				}
 			}
 		case *message.TextElement:
-			m = MSG{
+			m = global.MSG{
 				"type": "text",
 				"data": map[string]string{"text": o.Content},
 			}
 		case *message.LightAppElement:
-			// m = MSG{
-			// 	"type": "text",
-			// 	"data": map[string]string{"text": o.Content},
-			// }
-			m = MSG{
+			m = global.MSG{
 				"type": "json",
 				"data": map[string]string{"data": o.Content},
 			}
 		case *message.AtElement:
 			if o.Target == 0 {
-				m = MSG{
+				m = global.MSG{
 					"type": "at",
 					"data": map[string]string{"qq": "all"},
 				}
 			} else {
-				m = MSG{
+				m = global.MSG{
 					"type": "at",
-					"data": map[string]string{"qq": fmt.Sprint(o.Target)},
+					"data": map[string]string{"qq": strconv.FormatUint(uint64(o.Target), 10)},
 				}
 			}
 		case *message.RedBagElement:
-			m = MSG{
+			m = global.MSG{
 				"type": "redbag",
 				"data": map[string]string{"title": o.Title},
 			}
 		case *message.ForwardElement:
-			m = MSG{
+			m = global.MSG{
 				"type": "forward",
 				"data": map[string]string{"id": o.ResId},
 			}
 		case *message.FaceElement:
-			m = MSG{
+			m = global.MSG{
 				"type": "face",
-				"data": map[string]string{"id": fmt.Sprint(o.Index)},
+				"data": map[string]string{"id": strconv.FormatInt(int64(o.Index), 10)},
 			}
 		case *message.VoiceElement:
-			if ur {
-				m = MSG{
-					"type": "record",
-					"data": map[string]string{"file": o.Name},
-				}
-			} else {
-				m = MSG{
-					"type": "record",
-					"data": map[string]string{"file": o.Name, "url": o.Url},
-				}
+			m = global.MSG{
+				"type": "record",
+				"data": map[string]string{"file": o.Name, "url": o.Url},
 			}
 		case *message.ShortVideoElement:
-			if ur {
-				m = MSG{
-					"type": "video",
-					"data": map[string]string{"file": o.Name},
-				}
-			} else {
-				m = MSG{
-					"type": "video",
-					"data": map[string]string{"file": o.Name, "url": o.Url},
-				}
-			}
-		case *message.ImageElement:
-			if ur {
-				m = MSG{
-					"type": "image",
-					"data": map[string]string{"file": o.Filename},
-				}
-			} else {
-				m = MSG{
-					"type": "image",
-					"data": map[string]string{"file": o.Filename, "url": o.Url},
-				}
+			m = global.MSG{
+				"type": "video",
+				"data": map[string]string{"file": o.Name, "url": o.Url},
 			}
 		case *message.GroupImageElement:
-			if ur {
-				m = MSG{
-					"type": "image",
-					"data": map[string]string{"file": hex.EncodeToString(o.Md5) + ".image"},
-				}
-			} else {
-				m = MSG{
-					"type": "image",
-					"data": map[string]string{"file": hex.EncodeToString(o.Md5) + ".image", "url": o.Url},
-				}
+			data := map[string]string{"file": hex.EncodeToString(o.Md5) + ".image", "url": o.Url, "subType": strconv.FormatInt(int64(o.ImageBizType), 10)}
+			switch {
+			case o.Flash:
+				data["type"] = "flash"
+			case o.EffectID != 0:
+				data["type"] = "show"
+				data["id"] = strconv.FormatInt(int64(o.EffectID), 10)
+			}
+			m = global.MSG{
+				"type": "image",
+				"data": data,
+			}
+		case *message.GuildImageElement:
+			data := map[string]string{"file": hex.EncodeToString(o.Md5) + ".image", "url": o.Url}
+			m = global.MSG{
+				"type": "image",
+				"data": data,
 			}
 		case *message.FriendImageElement:
-			if ur {
-				m = MSG{
-					"type": "image",
-					"data": map[string]string{"file": hex.EncodeToString(o.Md5) + ".image"},
-				}
-			} else {
-				m = MSG{
-					"type": "image",
-					"data": map[string]string{"file": hex.EncodeToString(o.Md5) + ".image", "url": o.Url},
-				}
+			data := map[string]string{"file": hex.EncodeToString(o.Md5) + ".image", "url": o.Url}
+			if o.Flash {
+				data["type"] = "flash"
 			}
-		case *message.GroupFlashImgElement:
-			return []MSG{{
+			m = global.MSG{
 				"type": "image",
-				"data": map[string]string{"file": o.Filename, "type": "flash"},
-			}}
-		case *message.FriendFlashImgElement:
-			return []MSG{{
-				"type": "image",
-				"data": map[string]string{"file": o.Filename, "type": "flash"},
-			}}
+				"data": data,
+			}
+		case *message.DiceElement:
+			m = global.MSG{
+				"type": "dice",
+				"data": map[string]string{"value": fmt.Sprint(o.Value)},
+			}
+		case *message.MarketFaceElement:
+			m = global.MSG{
+				"type": "text",
+				"data": map[string]string{"text": o.Name},
+			}
 		case *message.ServiceElement:
 			if isOk := strings.Contains(o.Content, "<?xml"); isOk {
-				m = MSG{
+				m = global.MSG{
 					"type": "xml",
-					"data": map[string]string{"data": o.Content, "resid": fmt.Sprintf("%d", o.Id)},
+					"data": map[string]string{"data": o.Content, "resid": strconv.FormatInt(int64(o.Id), 10)},
 				}
 			} else {
-				m = MSG{
+				m = global.MSG{
 					"type": "json",
-					"data": map[string]string{"data": o.Content, "resid": fmt.Sprintf("%d", o.Id)},
+					"data": map[string]string{"data": o.Content, "resid": strconv.FormatInt(int64(o.Id), 10)},
 				}
+			}
+		case *message.AnimatedSticker:
+			m = global.MSG{
+				"type": "face",
+				"data": map[string]string{"id": strconv.FormatInt(int64(o.ID), 10), "type": "sticker"},
 			}
 		default:
 			continue
@@ -303,7 +261,12 @@ func ToArrayMessage(e []message.IMessageElement, id int64, isRaw ...bool) (r []M
 }
 
 // ToStringMessage 将消息元素数组转为字符串以用于消息上报
-func ToStringMessage(e []message.IMessageElement, id int64, isRaw ...bool) (r string) {
+func ToStringMessage(e []message.IMessageElement, source MessageSource, isRaw ...bool) (r string) {
+	sb := global.NewBuffer()
+	sb.Reset()
+	write := func(format string, a ...interface{}) {
+		_, _ = fmt.Fprintf(sb, format, a...)
+	}
 	ur := false
 	if len(isRaw) != 0 {
 		ur = isRaw[0]
@@ -314,95 +277,227 @@ func ToStringMessage(e []message.IMessageElement, id int64, isRaw ...bool) (r st
 		_, ok := e.(*message.ReplyElement)
 		return ok
 	})
-	if reply != nil {
+	if reply != nil && source.SourceType&(MessageSourceGroup|MessageSourcePrivate) != 0 {
 		replyElem := reply.(*message.ReplyElement)
-		if ExtraReplyData {
-			r += fmt.Sprintf("[CQ:reply,id=%d,seq=%d,qq=%d,time=%d,text=%s]",
-				toGlobalID(id, replyElem.ReplySeq),
+		rid := int64(source.PrimaryID)
+		if rid == 0 {
+			rid = replyElem.Sender
+		}
+		if replyElem.GroupID != 0 {
+			rid = replyElem.GroupID
+		}
+		if base.ExtraReplyData {
+			write("[CQ:reply,id=%d,seq=%d,qq=%d,time=%d,text=%s]",
+				db.ToGlobalID(rid, replyElem.ReplySeq),
 				replyElem.ReplySeq, replyElem.Sender, replyElem.Time,
-				CQCodeEscapeValue(CQCodeEscapeText(ToStringMessage(replyElem.Elements, id))))
+				CQCodeEscapeValue(ToStringMessage(replyElem.Elements, source)))
 		} else {
-			r += fmt.Sprintf("[CQ:reply,id=%d]", toGlobalID(id, replyElem.ReplySeq))
+			write("[CQ:reply,id=%d]", db.ToGlobalID(rid, replyElem.ReplySeq))
 		}
 	}
 	for i, elem := range e {
 		switch o := elem.(type) {
 		case *message.ReplyElement:
-			if RemoveReplyAt && len(e) > i+1 {
+			if base.RemoveReplyAt && len(e) > i+1 {
 				elem, ok := e[i+1].(*message.AtElement)
 				if ok && elem.Target == o.Sender {
 					e[i+1] = nil
 				}
 			}
 		case *message.TextElement:
-			r += CQCodeEscapeText(o.Content)
+			sb.WriteString(CQCodeEscapeText(o.Content))
 		case *message.AtElement:
 			if o.Target == 0 {
-				r += "[CQ:at,qq=all]"
+				write("[CQ:at,qq=all]")
 				continue
 			}
-			r += fmt.Sprintf("[CQ:at,qq=%d]", o.Target)
+			write("[CQ:at,qq=%d]", uint64(o.Target))
 		case *message.RedBagElement:
-			r += fmt.Sprintf("[CQ:redbag,title=%s]", o.Title)
+			write("[CQ:redbag,title=%s]", o.Title)
 		case *message.ForwardElement:
-			r += fmt.Sprintf("[CQ:forward,id=%s]", o.ResId)
+			write("[CQ:forward,id=%s]", o.ResId)
 		case *message.FaceElement:
-			r += fmt.Sprintf(`[CQ:face,id=%d]`, o.Index)
+			write(`[CQ:face,id=%d]`, o.Index)
 		case *message.VoiceElement:
 			if ur {
-				r += fmt.Sprintf(`[CQ:record,file=%s]`, o.Name)
+				write(`[CQ:record,file=%s]`, o.Name)
 			} else {
-				r += fmt.Sprintf(`[CQ:record,file=%s,url=%s]`, o.Name, CQCodeEscapeValue(o.Url))
+				write(`[CQ:record,file=%s,url=%s]`, o.Name, CQCodeEscapeValue(o.Url))
 			}
 		case *message.ShortVideoElement:
 			if ur {
-				r += fmt.Sprintf(`[CQ:video,file=%s]`, o.Name)
+				write(`[CQ:video,file=%s]`, o.Name)
 			} else {
-				r += fmt.Sprintf(`[CQ:video,file=%s,url=%s]`, o.Name, CQCodeEscapeValue(o.Url))
-			}
-		case *message.ImageElement:
-			if ur {
-				r += fmt.Sprintf(`[CQ:image,file=%s]`, o.Filename)
-			} else {
-				r += fmt.Sprintf(`[CQ:image,file=%s,url=%s]`, o.Filename, CQCodeEscapeValue(o.Url))
+				write(`[CQ:video,file=%s,url=%s]`, o.Name, CQCodeEscapeValue(o.Url))
 			}
 		case *message.GroupImageElement:
+			var arg string
+			if o.Flash {
+				arg = ",type=flash"
+			} else if o.EffectID != 0 {
+				arg = ",type=show,id=" + strconv.FormatInt(int64(o.EffectID), 10)
+			}
+			arg += ",subType=" + strconv.FormatInt(int64(o.ImageBizType), 10)
 			if ur {
-				r += fmt.Sprintf("[CQ:image,file=%s]", hex.EncodeToString(o.Md5)+".image")
+				write("[CQ:image,file=%s%s]", hex.EncodeToString(o.Md5)+".image", arg)
 			} else {
-				r += fmt.Sprintf("[CQ:image,file=%s,url=%s]", hex.EncodeToString(o.Md5)+".image", CQCodeEscapeValue(o.Url))
+				write("[CQ:image,file=%s,url=%s%s]", hex.EncodeToString(o.Md5)+".image", CQCodeEscapeValue(o.Url), arg)
 			}
 		case *message.FriendImageElement:
-			if ur {
-				r += fmt.Sprintf("[CQ:image,file=%s]", hex.EncodeToString(o.Md5)+".image")
-			} else {
-				r += fmt.Sprintf("[CQ:image,file=%s,url=%s]", hex.EncodeToString(o.Md5)+".image", CQCodeEscapeValue(o.Url))
+			var arg string
+			if o.Flash {
+				arg = ",type=flash"
 			}
-		case *message.GroupFlashImgElement:
-			return fmt.Sprintf("[CQ:image,type=flash,file=%s]", o.Filename)
-		case *message.FriendFlashImgElement:
-			return fmt.Sprintf("[CQ:image,type=flash,file=%s]", o.Filename)
+			if ur {
+				write("[CQ:image,file=%s%s]", hex.EncodeToString(o.Md5)+".image", arg)
+			} else {
+				write("[CQ:image,file=%s,url=%s%s]", hex.EncodeToString(o.Md5)+".image", CQCodeEscapeValue(o.Url), arg)
+			}
+		case *message.GuildImageElement:
+			write("[CQ:image,file=%s,url=%s]", hex.EncodeToString(o.Md5)+".image", CQCodeEscapeValue(o.Url))
+		case *message.DiceElement:
+			write("[CQ:dice,value=%v]", o.Value)
+		case *message.MarketFaceElement:
+			sb.WriteString(o.Name)
 		case *message.ServiceElement:
 			if isOk := strings.Contains(o.Content, "<?xml"); isOk {
-				r += fmt.Sprintf(`[CQ:xml,data=%s,resid=%d]`, CQCodeEscapeValue(o.Content), o.Id)
+				write(`[CQ:xml,data=%s,resid=%d]`, CQCodeEscapeValue(o.Content), o.Id)
 			} else {
-				r += fmt.Sprintf(`[CQ:json,data=%s,resid=%d]`, CQCodeEscapeValue(o.Content), o.Id)
+				write(`[CQ:json,data=%s,resid=%d]`, CQCodeEscapeValue(o.Content), o.Id)
 			}
 		case *message.LightAppElement:
-			r += fmt.Sprintf(`[CQ:json,data=%s]`, CQCodeEscapeValue(o.Content))
-			// r += CQCodeEscapeText(o.Content)
+			write(`[CQ:json,data=%s]`, CQCodeEscapeValue(o.Content))
+		case *message.AnimatedSticker:
+			write(`[CQ:face,id=%d,type=sticker]`, o.ID)
+		}
+	}
+	r = sb.String() // 内部已拷贝
+	global.PutBuffer(sb)
+	return
+}
+
+// ToMessageContent 将消息转换成 Content. 忽略 Reply
+// 不同于 onebot 的 Array Message, 此函数转换出来的 Content 的 data 段为实际类型
+// 方便数据库查询
+func ToMessageContent(e []message.IMessageElement) (r []global.MSG) {
+	for _, elem := range e {
+		var m global.MSG
+		switch o := elem.(type) {
+		case *message.TextElement:
+			m = global.MSG{
+				"type": "text",
+				"data": global.MSG{"text": o.Content},
+			}
+		case *message.LightAppElement:
+			m = global.MSG{
+				"type": "json",
+				"data": global.MSG{"data": o.Content},
+			}
+		case *message.AtElement:
+			if o.Target == 0 {
+				m = global.MSG{
+					"type": "at",
+					"data": global.MSG{
+						"subType": "all",
+					},
+				}
+			} else {
+				m = global.MSG{
+					"type": "at",
+					"data": global.MSG{
+						"subType": "user",
+						"target":  o.Target,
+						"display": o.Display,
+					},
+				}
+			}
+		case *message.RedBagElement:
+			m = global.MSG{
+				"type": "redbag",
+				"data": global.MSG{"title": o.Title, "type": int(o.MsgType)},
+			}
+		case *message.ForwardElement:
+			m = global.MSG{
+				"type": "forward",
+				"data": global.MSG{"id": o.ResId},
+			}
+		case *message.FaceElement:
+			m = global.MSG{
+				"type": "face",
+				"data": global.MSG{"id": o.Index},
+			}
+		case *message.VoiceElement:
+			m = global.MSG{
+				"type": "record",
+				"data": global.MSG{"file": o.Name, "url": o.Url},
+			}
+		case *message.ShortVideoElement:
+			m = global.MSG{
+				"type": "video",
+				"data": global.MSG{"file": o.Name, "url": o.Url},
+			}
+		case *message.GroupImageElement:
+			data := global.MSG{"file": hex.EncodeToString(o.Md5) + ".image", "url": o.Url, "subType": uint32(o.ImageBizType)}
+			switch {
+			case o.Flash:
+				data["type"] = "flash"
+			case o.EffectID != 0:
+				data["type"] = "show"
+				data["id"] = o.EffectID
+			}
+			m = global.MSG{
+				"type": "image",
+				"data": data,
+			}
+		case *message.GuildImageElement:
+			m = global.MSG{
+				"type": "image",
+				"data": global.MSG{"file": hex.EncodeToString(o.Md5) + ".image", "url": o.Url},
+			}
+		case *message.FriendImageElement:
+			data := global.MSG{"file": hex.EncodeToString(o.Md5) + ".image", "url": o.Url}
+			if o.Flash {
+				data["type"] = "flash"
+			}
+			m = global.MSG{
+				"type": "image",
+				"data": data,
+			}
+		case *message.DiceElement:
+			m = global.MSG{"type": "dice", "data": global.MSG{"value": o.Value}}
+		case *message.MarketFaceElement:
+			m = global.MSG{"type": "text", "data": global.MSG{"text": o.Name}}
+		case *message.ServiceElement:
+			if isOk := strings.Contains(o.Content, "<?xml"); isOk {
+				m = global.MSG{
+					"type": "xml",
+					"data": global.MSG{"data": o.Content, "resid": o.Id},
+				}
+			} else {
+				m = global.MSG{
+					"type": "json",
+					"data": global.MSG{"data": o.Content, "resid": o.Id},
+				}
+			}
+		case *message.AnimatedSticker:
+			m = global.MSG{
+				"type": "face",
+				"data": global.MSG{"id": o.ID, "type": "sticker"},
+			}
+		default:
+			continue
+		}
+		if m != nil {
+			r = append(r, m)
 		}
 	}
 	return
 }
 
 // ConvertStringMessage 将消息字符串转为消息元素数组
-func (bot *CQBot) ConvertStringMessage(s string, isGroup bool) (r []message.IMessageElement) {
+func (bot *CQBot) ConvertStringMessage(raw string, sourceType MessageSourceType) (r []message.IMessageElement) {
 	var t, key string
 	d := map[string]string{}
-	ptr := unsafe.Pointer((*reflect.StringHeader)(unsafe.Pointer(&s)).Data)
-	l := len(s)
-	i, j, CQBegin := 0, 0, 0
 
 	saveCQCode := func() {
 		if t == "reply" { // reply 特殊处理
@@ -417,7 +512,7 @@ func (bot *CQBot) ConvertStringMessage(s string, isGroup bool) (r []message.IMes
 			switch {
 			case customText != "":
 				var elem *message.ReplyElement
-				var org MSG
+				var org db.StoredMessage
 				sender, senderErr := strconv.ParseInt(d["qq"], 10, 64)
 				if senderErr != nil && err != nil {
 					log.Warnf("警告: 自定义 Reply 元素中必须包含 Uin 或 id")
@@ -429,14 +524,14 @@ func (bot *CQBot) ConvertStringMessage(s string, isGroup bool) (r []message.IMes
 				}
 				messageSeq, seqErr := strconv.ParseInt(d["seq"], 10, 64)
 				if err == nil {
-					org = bot.GetMessage(int32(mid))
+					org, _ = db.GetMessageByGlobalID(int32(mid))
 				}
 				if org != nil {
 					elem = &message.ReplyElement{
-						ReplySeq: org["message-id"].(int32),
-						Sender:   org["sender"].(message.Sender).Uin,
-						Time:     org["time"].(int32),
-						Elements: bot.ConvertStringMessage(customText, isGroup),
+						ReplySeq: org.GetAttribute().MessageSeq,
+						Sender:   org.GetAttribute().SenderUin,
+						Time:     int32(org.GetAttribute().Timestamp),
+						Elements: bot.ConvertStringMessage(customText, sourceType),
 					}
 					if senderErr != nil {
 						elem.Sender = sender
@@ -452,19 +547,19 @@ func (bot *CQBot) ConvertStringMessage(s string, isGroup bool) (r []message.IMes
 						ReplySeq: int32(messageSeq),
 						Sender:   sender,
 						Time:     int32(msgTime),
-						Elements: bot.ConvertStringMessage(customText, isGroup),
+						Elements: bot.ConvertStringMessage(customText, sourceType),
 					}
 				}
 				r = append([]message.IMessageElement{elem}, r...)
 			case err == nil:
-				org := bot.GetMessage(int32(mid))
-				if org != nil {
+				org, err := db.GetMessageByGlobalID(int32(mid))
+				if err == nil {
 					r = append([]message.IMessageElement{
 						&message.ReplyElement{
-							ReplySeq: org["message-id"].(int32),
-							Sender:   org["sender"].(message.Sender).Uin,
-							Time:     org["time"].(int32),
-							Elements: bot.ConvertStringMessage(org["message"].(string), isGroup),
+							ReplySeq: org.GetAttribute().MessageSeq,
+							Sender:   org.GetAttribute().SenderUin,
+							Time:     int32(org.GetAttribute().Timestamp),
+							Elements: bot.ConvertContentMessage(org.GetContent(), sourceType),
 						},
 					}, r...)
 				}
@@ -485,10 +580,14 @@ func (bot *CQBot) ConvertStringMessage(s string, isGroup bool) (r []message.IMes
 			}
 			return
 		}
-		elem, err := bot.ToElement(t, d, isGroup)
+		elem, err := bot.ToElement(t, d, sourceType)
 		if err != nil {
-			org := s[CQBegin:i]
-			if !IgnoreInvalidCQCode {
+			org := "[CQ:" + t
+			for k, v := range d {
+				org += "," + k + "=" + v
+			}
+			org += "]"
+			if !base.IgnoreInvalidCQCode {
 				log.Warnf("转换CQ码 %v 时出现错误: %v 将原样发送.", org, err)
 				r = append(r, message.NewText(org))
 			} else {
@@ -504,91 +603,76 @@ func (bot *CQBot) ConvertStringMessage(s string, isGroup bool) (r []message.IMes
 		}
 	}
 
-S1: // Plain Text
-	for ; i < l; i++ {
-		if *(*byte)(add(ptr, uintptr(i))) == '[' && i+4 < l &&
-			*(*uint32)(add(ptr, uintptr(i))) == magicCQ { // Magic :uint32([]byte("[CQ:"))
-			if i > j {
-				if SplitURL {
-					for _, str := range global.SplitURL(CQCodeUnescapeText(s[j:i])) {
-						r = append(r, message.NewText(str))
-					}
-				} else {
-					r = append(r, message.NewText(CQCodeUnescapeText(s[j:i])))
+	for raw != "" {
+		i := 0
+		for i < len(raw) && !(raw[i] == '[' && i+4 < len(raw) && raw[i:i+4] == "[CQ:") {
+			i++
+		}
+		if i > 0 {
+			if base.SplitURL {
+				for _, txt := range param.SplitURL(CQCodeUnescapeText(raw[:i])) {
+					r = append(r, message.NewText(txt))
 				}
+			} else {
+				r = append(r, message.NewText(CQCodeUnescapeText(raw[:i])))
 			}
-			CQBegin = i
-			i += 4
-			j = i
-			goto S2
 		}
-	}
-	goto End
-S2: // CQCode Type
-	for k := range d { // 内存复用，减小GC压力
-		delete(d, k)
-	}
-	for ; i < l; i++ {
-		switch *(*byte)(add(ptr, uintptr(i))) {
-		case ',': // CQ Code with params
-			t = s[j:i]
-			i++
-			j = i
-			goto S3
-		case ']': // CQ Code without params
-			t = s[j:i]
-			i++
-			j = i
-			saveCQCode()
-			goto S1
+
+		if i+4 > len(raw) {
+			return
 		}
-	}
-	goto End
-S3: // CQCode param key
-	for ; i < l; i++ {
-		if *(*byte)(add(ptr, uintptr(i))) == '=' {
-			key = s[j:i]
+		raw = raw[i+4:] // skip "[CQ:"
+		i = 0
+		for i < len(raw) && raw[i] != ',' && raw[i] != ']' {
 			i++
-			j = i
-			goto S4
 		}
-	}
-	goto End
-S4: // CQCode param value
-	for ; i < l; i++ {
-		switch *(*byte)(add(ptr, uintptr(i))) {
-		case ',': // more param
-			d[key] = CQCodeUnescapeValue(s[j:i])
-			i++
-			j = i
-			goto S3
-		case ']':
-			d[key] = CQCodeUnescapeValue(s[j:i])
-			i++
-			j = i
-			saveCQCode()
-			goto S1
+		if i+1 > len(raw) {
+			return
 		}
-	}
-	goto End
-End:
-	if i > j {
-		if SplitURL {
-			for _, str := range global.SplitURL(CQCodeUnescapeText(s[j:i])) {
-				r = append(r, message.NewText(str))
+		t = raw[:i]
+		for k := range d { // clear the map, reuse it
+			delete(d, k)
+		}
+		raw = raw[i:]
+		i = 0
+		for {
+			if raw[0] == ']' {
+				saveCQCode()
+				raw = raw[1:]
+				break
 			}
-		} else {
-			r = append(r, message.NewText(CQCodeUnescapeText(s[j:i])))
+			raw = raw[1:]
+
+			for i < len(raw) && raw[i] != '=' {
+				i++
+			}
+			if i+1 > len(raw) {
+				return
+			}
+			key = raw[:i]
+			raw = raw[i+1:] // skip "="
+			i = 0
+			for i < len(raw) && raw[i] != ',' && raw[i] != ']' {
+				i++
+			}
+
+			if i+1 > len(raw) {
+				return
+			}
+			d[key] = CQCodeUnescapeValue(raw[:i])
+			raw = raw[i:]
+			i = 0
 		}
 	}
 	return
 }
 
 // ConvertObjectMessage 将消息JSON对象转为消息元素数组
-func (bot *CQBot) ConvertObjectMessage(m gjson.Result, isGroup bool) (r []message.IMessageElement) {
+func (bot *CQBot) ConvertObjectMessage(m gjson.Result, sourceType MessageSourceType) (r []message.IMessageElement) {
+	d := make(map[string]string)
 	convertElem := func(e gjson.Result) {
 		t := e.Get("type").Str
-		if t == "reply" && isGroup {
+		if t == "reply" && sourceType&(MessageSourceGroup|MessageSourcePrivate) != 0 {
 			if len(r) > 0 {
 				if _, ok := r[0].(*message.ReplyElement); ok {
 					log.Warnf("警告: 一条信息只能包含一个 Reply 元素.")
@@ -600,10 +684,10 @@ func (bot *CQBot) ConvertObjectMessage(m gjson.Result, isGroup bool) (r []messag
 			switch {
 			case customText != "":
 				var elem *message.ReplyElement
-				var org MSG
-				sender, senderErr := strconv.ParseInt(e.Get("data.qq").String(), 10, 64)
+				var org db.StoredMessage
+				sender, senderErr := strconv.ParseInt(e.Get("data.[user_id,qq]").String(), 10, 64)
 				if senderErr != nil && err != nil {
-					log.Warnf("警告: 自定义 Reply 元素中必须包含 Uin 或 id")
+					log.Warnf("警告: 自定义 Reply 元素中必须包含 user_id 或 id")
 					break
 				}
 				msgTime, timeErr := strconv.ParseInt(e.Get("data.time").String(), 10, 64)
@@ -612,14 +696,14 @@ func (bot *CQBot) ConvertObjectMessage(m gjson.Result, isGroup bool) (r []messag
 				}
 				messageSeq, seqErr := strconv.ParseInt(e.Get("data.seq").String(), 10, 64)
 				if err == nil {
-					org = bot.GetMessage(int32(mid))
+					org, _ = db.GetMessageByGlobalID(int32(mid))
 				}
 				if org != nil {
 					elem = &message.ReplyElement{
-						ReplySeq: org["message-id"].(int32),
-						Sender:   org["sender"].(message.Sender).Uin,
-						Time:     org["time"].(int32),
-						Elements: bot.ConvertStringMessage(customText, isGroup),
+						ReplySeq: org.GetAttribute().MessageSeq,
+						Sender:   org.GetAttribute().SenderUin,
+						Time:     int32(org.GetAttribute().Timestamp),
+						Elements: bot.ConvertStringMessage(customText, sourceType),
 					}
 					if senderErr != nil {
 						elem.Sender = sender
@@ -635,19 +719,19 @@ func (bot *CQBot) ConvertObjectMessage(m gjson.Result, isGroup bool) (r []messag
 						ReplySeq: int32(messageSeq),
 						Sender:   sender,
 						Time:     int32(msgTime),
-						Elements: bot.ConvertStringMessage(customText, isGroup),
+						Elements: bot.ConvertStringMessage(customText, sourceType),
 					}
 				}
 				r = append([]message.IMessageElement{elem}, r...)
 			case err == nil:
-				org := bot.GetMessage(int32(mid))
-				if org != nil {
+				org, err := db.GetMessageByGlobalID(int32(mid))
+				if err == nil {
 					r = append([]message.IMessageElement{
 						&message.ReplyElement{
-							ReplySeq: org["message-id"].(int32),
-							Sender:   org["sender"].(message.Sender).Uin,
-							Time:     org["time"].(int32),
-							Elements: bot.ConvertStringMessage(org["message"].(string), isGroup),
+							ReplySeq: org.GetAttribute().MessageSeq,
+							Sender:   org.GetAttribute().SenderUin,
+							Time:     int32(org.GetAttribute().Timestamp),
+							Elements: bot.ConvertContentMessage(org.GetContent(), sourceType),
 						},
 					}, r...)
 				}
@@ -669,12 +753,14 @@ func (bot *CQBot) ConvertObjectMessage(m gjson.Result, isGroup bool) (r []messag
 			}
 			return
 		}
-		d := make(map[string]string)
+		for i := range d {
+			delete(d, i)
+		}
 		e.Get("data").ForEach(func(key, value gjson.Result) bool {
 			d[key.Str] = value.String()
 			return true
 		})
-		elem, err := bot.ToElement(t, d, isGroup)
+		elem, err := bot.ToElement(t, d, sourceType)
 		if err != nil {
 			log.Warnf("转换CQ码 (%v) 到MiraiGo Element时出现错误: %v 将忽略本段CQ码.", e.Raw, err)
 			return
@@ -687,15 +773,85 @@ func (bot *CQBot) ConvertObjectMessage(m gjson.Result, isGroup bool) (r []messag
 		}
 	}
 	if m.Type == gjson.String {
-		return bot.ConvertStringMessage(m.Str, isGroup)
+		return bot.ConvertStringMessage(m.Str, sourceType)
 	}
 	if m.IsArray() {
-		for _, e := range m.Array() {
+		m.ForEach(func(_, e gjson.Result) bool {
 			convertElem(e)
-		}
+			return true
+		})
 	}
 	if m.IsObject() {
 		convertElem(m)
+	}
+	return
+}
+
+// ConvertContentMessage 将数据库用的 content 转换为消息元素数组
+func (bot *CQBot) ConvertContentMessage(content []global.MSG, sourceType MessageSourceType) (r []message.IMessageElement) {
+	for _, c := range content {
+		data := c["data"].(global.MSG)
+		switch c["type"] {
+		case "text":
+			r = append(r, message.NewText(data["text"].(string)))
+		case "image":
+			e, err := bot.makeImageOrVideoElem(map[string]string{"file": data["file"].(string)}, false, sourceType)
+			if err != nil {
+				log.Warnf("make image elem error: %v", err)
+				continue
+			}
+			flash, id := false, int32(0)
+			if t, ok := data["type"]; ok {
+				if t.(string) == "flash" {
+					flash = true
+				}
+				if t.(string) == "show" {
+					id = data["id"].(int32)
+					if id < 40000 || id >= 40006 {
+						id = 40000
+					}
+				}
+			}
+			switch img := e.(type) {
+			case *LocalImageElement:
+				img.Flash = flash
+				img.EffectID = id
+			case *message.GroupImageElement:
+				img.Flash = flash
+				img.EffectID = id
+				img.ImageBizType = message.ImageBizType(data["subType"].(uint32))
+			case *message.FriendImageElement:
+				img.Flash = flash
+			}
+			r = append(r, e)
+		case "at":
+			switch data["subType"].(string) {
+			case "all":
+				r = append(r, message.NewAt(0))
+			case "user":
+				r = append(r, message.NewAt(data["target"].(int64), data["display"].(string)))
+			default:
+				continue
+			}
+		case "redbag":
+			r = append(r, &message.RedBagElement{
+				MsgType: message.RedBagMessageType(data["type"].(int)),
+				Title:   data["title"].(string),
+			})
+		case "forward":
+			r = append(r, &message.ForwardElement{
+				ResId: data["id"].(string),
+			})
+		case "face":
+			r = append(r, message.NewFace(data["id"].(int32)))
+		case "video":
+			e, err := bot.makeImageOrVideoElem(map[string]string{"file": data["file"].(string)}, true, sourceType)
+			if err != nil {
+				log.Warnf("make image elem error: %v", err)
+				continue
+			}
+			r = append(r, e)
+		}
 	}
 	return
 }
@@ -705,68 +861,51 @@ func (bot *CQBot) ConvertObjectMessage(m gjson.Result, isGroup bool) (r []messag
 // 返回 interface{} 存在三种类型
 //
 // message.IMessageElement []message.IMessageElement nil
-func (bot *CQBot) ToElement(t string, d map[string]string, isGroup bool) (m interface{}, err error) {
+func (bot *CQBot) ToElement(t string, d map[string]string, sourceType MessageSourceType) (m interface{}, err error) {
 	switch t {
 	case "text":
-		if SplitURL {
+		if base.SplitURL {
 			var ret []message.IMessageElement
-			for _, text := range global.SplitURL(d["text"]) {
+			for _, text := range param.SplitURL(d["text"]) {
 				ret = append(ret, message.NewText(text))
 			}
 			return ret, nil
 		}
 		return message.NewText(d["text"]), nil
 	case "image":
-		img, err := bot.makeImageOrVideoElem(d, false, isGroup)
+		img, err := bot.makeImageOrVideoElem(d, false, sourceType)
 		if err != nil {
 			return nil, err
 		}
 		tp := d["type"]
-		if tp != "show" && tp != "flash" {
-			return img, nil
-		}
-		if i, ok := img.(*LocalImageElement); ok { // 秀图，闪照什么的就直接传了吧
-			if isGroup {
-				img, err = bot.UploadLocalImageAsGroup(1, i)
-			} else {
-				img, err = bot.UploadLocalImageAsPrivate(1, i)
-			}
-			if err != nil {
-				return nil, err
-			}
-		}
+		flash, id := false, int64(0)
 		switch tp {
 		case "flash":
-			if i, ok := img.(*message.GroupImageElement); ok {
-				return &message.GroupFlashPicElement{GroupImageElement: *i}, nil
-			}
-			if i, ok := img.(*message.FriendImageElement); ok {
-				return &message.FriendFlashPicElement{FriendImageElement: *i}, nil
-			}
+			flash = true
 		case "show":
-			id, _ := strconv.ParseInt(d["id"], 10, 64)
+			id, _ = strconv.ParseInt(d["id"], 10, 64)
 			if id < 40000 || id >= 40006 {
 				id = 40000
 			}
-			if i, ok := img.(*message.GroupImageElement); ok {
-				return &message.GroupShowPicElement{GroupImageElement: *i, EffectId: int32(id)}, nil
-			}
-			return img, nil // 私聊还没做
+		default:
+			return img, err
 		}
-
+		switch img := img.(type) {
+		case *LocalImageElement:
+			img.Flash = flash
+			img.EffectID = int32(id)
+		case *message.GroupImageElement:
+			img.Flash = flash
+			img.EffectID = int32(id)
+			i, _ := strconv.ParseInt(d["subType"], 10, 64)
+			img.ImageBizType = message.ImageBizType(i)
+		case *message.FriendImageElement:
+			img.Flash = flash
+		}
+		return img, err
 	case "poke":
 		t, _ := strconv.ParseInt(d["qq"], 10, 64)
 		return &PokeElement{Target: t}, nil
-	case "gift":
-		if !isGroup {
-			return nil, errors.New("private gift unsupported") // no free private gift
-		}
-		t, _ := strconv.ParseInt(d["qq"], 10, 64)
-		id, _ := strconv.Atoi(d["id"])
-		if id < 0 || id >= 14 {
-			return nil, errors.New("invalid gift id")
-		}
-		return &GiftElement{Target: t, GiftID: GiftID[id]}, nil
 	case "tts":
 		defer func() {
 			if r := recover(); r != nil {
@@ -778,7 +917,7 @@ func (bot *CQBot) ToElement(t string, d map[string]string, isGroup bool) (m inte
 		if err != nil {
 			return nil, err
 		}
-		return &message.VoiceElement{Data: data}, nil
+		return &message.VoiceElement{Data: base.ResampleSilk(data)}, nil
 	case "record":
 		f := d["file"]
 		data, err := global.FindFile(f, d["cache"], global.VoicePath)
@@ -787,6 +926,12 @@ func (bot *CQBot) ToElement(t string, d map[string]string, isGroup bool) (m inte
 		}
 		if err != nil {
 			return nil, err
+		}
+		if !global.IsAMRorSILK(data) {
+			lawful, mt := base.IsLawfulAudio(bytes.NewReader(data))
+			if !lawful {
+				return nil, errors.New("audio type error: " + mt)
+			}
 		}
 		if !global.IsAMRorSILK(data) {
 			data, err = global.EncoderSilk(data)
@@ -799,6 +944,9 @@ func (bot *CQBot) ToElement(t string, d map[string]string, isGroup bool) (m inte
 		id, err := strconv.Atoi(d["id"])
 		if err != nil {
 			return nil, err
+		}
+		if d["type"] == "sticker" {
+			return &message.AnimatedSticker{ID: int32(id)}, nil
 		}
 		return message.NewFace(int32(id)), nil
 	case "at":
@@ -823,7 +971,6 @@ func (bot *CQBot) ToElement(t string, d map[string]string, isGroup bool) (m inte
 			if !info.Get("track_info").Exists() {
 				return nil, errors.New("song not found")
 			}
-			aid := strconv.FormatInt(info.Get("track_info.album.id").Int(), 10)
 			name := info.Get("track_info.name").Str
 			mid := info.Get("track_info.mid").Str
 			albumMid := info.Get("track_info.album.mid").Str
@@ -831,9 +978,6 @@ func (bot *CQBot) ToElement(t string, d map[string]string, isGroup bool) (m inte
 			jumpURL := "https://i.y.qq.com/v8/playsong.html?platform=11&appshare=android_qq&appversion=10030010&hosteuin=oKnlNenz7i-s7c**&songmid=" + mid + "&type=0&appsongtype=1&_wv=1&source=qq&ADTAG=qfshare"
 			purl := gjson.ParseBytes(pinfo).Get("url_mid.data.midurlinfo.0.purl").Str
 			preview := "http://y.gtimg.cn/music/photo_new/T002R180x180M000" + albumMid + ".jpg"
-			if len(aid) < 2 {
-				return nil, errors.New("song error")
-			}
 			content := info.Get("track_info.singer.0.name").Str
 			if d["content"] != "" {
 				content = d["content"]
@@ -874,24 +1018,26 @@ func (bot *CQBot) ToElement(t string, d map[string]string, isGroup bool) (m inte
 		}
 		if d["type"] == "custom" {
 			if d["subtype"] != "" {
-				subtype := map[string]int{
-					"qq":    message.QQMusic,
-					"163":   message.CloudMusic,
-					"migu":  message.MiguMusic,
-					"kugou": message.KugouMusic,
-					"kuwo":  message.KuwoMusic,
-				}
-				musicType := 0
-				if tp, ok := subtype[d["subtype"]]; ok {
-					musicType = tp
+				var subType int
+				switch d["subtype"] {
+				default:
+					subType = message.QQMusic
+				case "163":
+					subType = message.CloudMusic
+				case "migu":
+					subType = message.MiguMusic
+				case "kugou":
+					subType = message.KugouMusic
+				case "kuwo":
+					subType = message.KuwoMusic
 				}
 				return &message.MusicShareElement{
-					MusicType:  musicType,
+					MusicType:  subType,
 					Title:      d["title"],
 					Summary:    d["content"],
 					Url:        d["url"],
 					PictureUrl: d["image"],
-					MusicUrl:   d["purl"],
+					MusicUrl:   d["audio"],
 				}, nil
 			}
 			xml := fmt.Sprintf(`<?xml version='1.0' encoding='UTF-8' standalone='yes' ?><msg serviceID="2" templateID="1" action="web" brief="[分享] %s" sourceMsgId="0" url="%s" flag="0" adverSign="0" multiMsgFlag="0"><item layout="2"><audio cover="%s" src="%s"/><title>%s</title><summary>%s</summary></item><source name="音乐" icon="https://i.gtimg.cn/open/app_icon/01/07/98/56/1101079856_100_m.png" url="http://web.p.qq.com/qqmpmobile/aio/app.html?id=1101079856" action="app" a_actionData="com.tencent.qqmusic" i_actionData="tencent1101079856://" appid="1101079856" /></msg>`,
@@ -903,6 +1049,13 @@ func (bot *CQBot) ToElement(t string, d map[string]string, isGroup bool) (m inte
 			}, nil
 		}
 		return nil, errors.New("unsupported music type: " + d["type"])
+	case "dice":
+		value := d["value"]
+		i, _ := strconv.ParseInt(value, 10, 64)
+		if i < 0 || i > 6 {
+			return nil, errors.New("invalid dice value " + value)
+		}
+		return message.NewDice(int32(i)), nil
 	case "xml":
 		resID := d["resid"]
 		template := CQCodeEscapeValue(d["data"])
@@ -924,46 +1077,40 @@ func (bot *CQBot) ToElement(t string, d map[string]string, isGroup bool) (m inte
 		source := d["source"]
 		icon := d["icon"]
 		brief := d["brief"]
-		minWidth, _ := strconv.ParseInt(d["minwidth"], 10, 64)
-		if minWidth == 0 {
-			minWidth = 200
+		parseIntWithDefault := func(name string, origin int64) int64 {
+			v, _ := strconv.ParseInt(d[name], 10, 64)
+			if v <= 0 {
+				return origin
+			}
+			return v
 		}
-		minHeight, _ := strconv.ParseInt(d["minheight"], 10, 64)
-		if minHeight == 0 {
-			minHeight = 200
-		}
-		maxWidth, _ := strconv.ParseInt(d["maxwidth"], 10, 64)
-		if maxWidth == 0 {
-			maxWidth = 500
-		}
-		maxHeight, _ := strconv.ParseInt(d["maxheight"], 10, 64)
-		if maxHeight == 0 {
-			maxHeight = 1000
-		}
-		img, err := bot.makeImageOrVideoElem(d, false, isGroup)
+		minWidth := parseIntWithDefault("minwidth", 200)
+		maxWidth := parseIntWithDefault("maxwidth", 500)
+		minHeight := parseIntWithDefault("minheight", 200)
+		maxHeight := parseIntWithDefault("maxheight", 1000)
+		img, err := bot.makeImageOrVideoElem(d, false, sourceType)
 		if err != nil {
 			return nil, errors.New("send cardimage faild")
 		}
-		return bot.makeShowPic(img, source, brief, icon, minWidth, minHeight, maxWidth, maxHeight, isGroup)
+		return bot.makeShowPic(img, source, brief, icon, minWidth, minHeight, maxWidth, maxHeight, sourceType == MessageSourceGroup)
 	case "video":
-		cache := d["cache"]
-		if cache == "" {
-			cache = "1"
-		}
-		file, err := bot.makeImageOrVideoElem(d, true, isGroup)
+		file, err := bot.makeImageOrVideoElem(d, true, sourceType)
 		if err != nil {
 			return nil, err
 		}
-		v := file.(*LocalVideoElement)
+		v, ok := file.(*LocalVideoElement)
+		if !ok {
+			return file, nil
+		}
 		if v.File == "" {
 			return v, nil
 		}
 		var data []byte
 		if cover, ok := d["cover"]; ok {
-			data, _ = global.FindFile(cover, cache, global.ImagePath)
+			data, _ = global.FindFile(cover, d["cache"], global.ImagePath)
 		} else {
 			_ = global.ExtractCover(v.File, v.File+".jpg")
-			data, _ = ioutil.ReadFile(v.File + ".jpg")
+			data, _ = os.ReadFile(v.File + ".jpg")
 		}
 		v.thumb = bytes.NewReader(data)
 		video, _ := os.Open(v.File)
@@ -981,7 +1128,7 @@ func (bot *CQBot) ToElement(t string, d map[string]string, isGroup bool) (m inte
 			_, _ = video.Seek(0, io.SeekStart)
 			hash, _ := utils.ComputeMd5AndLength(video)
 			cacheFile := path.Join(global.CachePath, hex.EncodeToString(hash)+".mp4")
-			if global.PathExists(cacheFile) && cache == "1" {
+			if global.PathExists(cacheFile) && (d["cache"] == "" || d["cache"] == "1") {
 				goto ok
 			}
 			err = global.EncodeMP4(v.File, cacheFile)
@@ -995,13 +1142,13 @@ func (bot *CQBot) ToElement(t string, d map[string]string, isGroup bool) (m inte
 	default:
 		return nil, errors.New("unsupported cq code: " + t)
 	}
-	return nil, nil
 }
 
 // XMLEscape 将字符串c转义为XML字符串
 func XMLEscape(c string) string {
-	buf := new(bytes.Buffer)
-	_ = xml2.EscapeText(buf, []byte(c))
+	buf := global.NewBuffer()
+	defer global.PutBuffer(buf)
+	_ = xml2.EscapeText(buf, utils.S2B(c))
 	return buf.String()
 }
 
@@ -1014,12 +1161,39 @@ func XMLEscape(c string) string {
 ] -> &#93;
 
 */
-func CQCodeEscapeText(raw string) string {
-	ret := raw
-	ret = strings.ReplaceAll(ret, "&", "&amp;")
-	ret = strings.ReplaceAll(ret, "[", "&#91;")
-	ret = strings.ReplaceAll(ret, "]", "&#93;")
-	return ret
+func CQCodeEscapeText(s string) string {
+	count := strings.Count(s, "&")
+	count += strings.Count(s, "[")
+	count += strings.Count(s, "]")
+	if count == 0 {
+		return s
+	}
+
+	// Apply replacements to buffer.
+	var b strings.Builder
+	b.Grow(len(s) + count*4)
+	start := 0
+	for i := 0; i < count; i++ {
+		j := start
+		for index, r := range s[start:] {
+			if r == '&' || r == '[' || r == ']' {
+				j += index
+				break
+			}
+		}
+		b.WriteString(s[start:j])
+		switch s[j] {
+		case '&':
+			b.WriteString("&amp;")
+		case '[':
+			b.WriteString("&#91;")
+		case ']':
+			b.WriteString("&#93;")
+		}
+		start = j + 1
+	}
+	b.WriteString(s[start:])
+	return b.String()
 }
 
 /*CQCodeEscapeValue 将字符串value中部分字符转义
@@ -1074,33 +1248,27 @@ func CQCodeUnescapeValue(content string) string {
 }
 
 // makeImageOrVideoElem 图片 elem 生成器，单独拎出来，用于公用
-func (bot *CQBot) makeImageOrVideoElem(d map[string]string, video, group bool) (message.IMessageElement, error) {
+func (bot *CQBot) makeImageOrVideoElem(d map[string]string, video bool, sourceType MessageSourceType) (message.IMessageElement, error) {
 	f := d["file"]
-	if strings.HasPrefix(f, "http") || strings.HasPrefix(f, "https") {
-		cache := d["cache"]
-		c := d["c"]
-		if cache == "" {
-			cache = "1"
-		}
+	if strings.HasPrefix(f, "http") {
 		hash := md5.Sum([]byte(f))
 		cacheFile := path.Join(global.CachePath, hex.EncodeToString(hash[:])+".cache")
-		maxSize := func() int64 {
-			if video {
-				return maxVideoSize
-			}
-			return maxImageSize
-		}()
-		thread, _ := strconv.Atoi(c)
-		if global.PathExists(cacheFile) && cache == "1" {
-			goto hasCacheFile
+		maxSize := int64(maxImageSize)
+		if video {
+			maxSize = maxVideoSize
 		}
-		if global.PathExists(cacheFile) {
+		thread, _ := strconv.Atoi(d["c"])
+		exist := global.PathExists(cacheFile)
+		if exist && (d["cache"] == "" || d["cache"] == "1") {
+			goto useCacheFile
+		}
+		if exist {
 			_ = os.Remove(cacheFile)
 		}
 		if err := global.DownloadFileMultiThreading(f, cacheFile, maxSize, thread, nil); err != nil {
 			return nil, err
 		}
-	hasCacheFile:
+	useCacheFile:
 		if video {
 			return &LocalVideoElement{File: cacheFile}, nil
 		}
@@ -1111,7 +1279,7 @@ func (bot *CQBot) makeImageOrVideoElem(d map[string]string, video, group bool) (
 		if err != nil {
 			return nil, err
 		}
-		if strings.HasPrefix(fu.Path, "/") && runtime.GOOS == `windows` {
+		if runtime.GOOS == `windows` && strings.HasPrefix(fu.Path, "/") {
 			fu.Path = fu.Path[1:]
 		}
 		info, err := os.Stat(fu.Path)
@@ -1132,105 +1300,123 @@ func (bot *CQBot) makeImageOrVideoElem(d map[string]string, video, group bool) (
 		}
 		return &LocalImageElement{File: fu.Path}, nil
 	}
-	rawPath := path.Join(global.ImagePath, f)
-	if video {
-		rawPath = path.Join(global.VideoPath, f)
-		if !global.PathExists(rawPath) {
-			return nil, errors.New("invalid video")
-		}
-		if path.Ext(rawPath) == ".video" {
-			b, _ := ioutil.ReadFile(rawPath)
-			r := binary.NewReader(b)
-			return &LocalVideoElement{ShortVideoElement: message.ShortVideoElement{ // todo 检查缓存是否有效
-				Md5:       r.ReadBytes(16),
-				ThumbMd5:  r.ReadBytes(16),
-				Size:      r.ReadInt32(),
-				ThumbSize: r.ReadInt32(),
-				Name:      r.ReadString(),
-				Uuid:      r.ReadAvailable(),
-			}}, nil
-		}
-		return &LocalVideoElement{File: rawPath}, nil
-	}
-	if strings.HasPrefix(f, "base64") {
-		b, err := base64.StdEncoding.DecodeString(strings.ReplaceAll(f, "base64://", ""))
+	if strings.HasPrefix(f, "base64") && !video {
+		b, err := param.Base64DecodeString(strings.TrimPrefix(f, "base64://"))
 		if err != nil {
 			return nil, err
 		}
 		return &LocalImageElement{Stream: bytes.NewReader(b)}, nil
 	}
-	if !global.PathExists(rawPath) && global.PathExists(path.Join(global.ImagePathOld, f)) {
-		rawPath = path.Join(global.ImagePathOld, f)
-	}
-	if !global.PathExists(rawPath) && global.PathExists(rawPath+".cqimg") {
-		rawPath += ".cqimg"
-	}
-	if !global.PathExists(rawPath) && d["url"] != "" {
-		return bot.makeImageOrVideoElem(map[string]string{"file": d["url"]}, false, group)
-	}
-	if global.PathExists(rawPath) {
-		file, err := os.Open(rawPath)
-		if err != nil {
-			return nil, err
-		}
-		if path.Ext(rawPath) != ".image" && path.Ext(rawPath) != ".cqimg" {
-			return &LocalImageElement{Stream: file}, nil
-		}
-		b, err := ioutil.ReadAll(file)
-		if err != nil {
-			return nil, err
-		}
-		if len(b) < 20 {
-			return nil, errors.New("invalid local file")
-		}
-		var (
-			size int32
-			hash []byte
-			url  string
-		)
-		if path.Ext(rawPath) == ".cqimg" {
-			for _, line := range strings.Split(global.ReadAllText(rawPath), "\n") {
-				kv := strings.SplitN(line, "=", 2)
-				switch kv[0] {
-				case "md5":
-					hash, _ = hex.DecodeString(strings.ReplaceAll(kv[1], "\r", ""))
-				case "size":
-					t, _ := strconv.Atoi(strings.ReplaceAll(kv[1], "\r", ""))
-					size = int32(t)
+	rawPath := path.Join(global.ImagePath, f)
+	if video {
+		if strings.HasSuffix(f, ".video") {
+			hash, err := hex.DecodeString(strings.TrimSuffix(f, ".video"))
+			if err == nil {
+				if b := cache.Video.Get(hash); b != nil {
+					return bot.readVideoCache(b), nil
 				}
 			}
-		} else {
-			r := binary.NewReader(b)
-			hash = r.ReadBytes(16)
-			size = r.ReadInt32()
-			r.ReadString()
-			url = r.ReadString()
 		}
-		if size == 0 {
-			if url != "" {
-				return bot.makeImageOrVideoElem(map[string]string{"file": url}, false, group)
+		rawPath = path.Join(global.VideoPath, f)
+		if !global.PathExists(rawPath) {
+			return nil, errors.New("invalid video")
+		}
+		if path.Ext(rawPath) != ".video" {
+			return &LocalVideoElement{File: rawPath}, nil
+		}
+		b, _ := os.ReadFile(rawPath)
+		return bot.readVideoCache(b), nil
+	}
+	// 目前频道内上传的图片均无法被查询到, 需要单独处理
+	if sourceType == MessageSourceGuildChannel {
+		cacheFile := path.Join(global.ImagePath, "guild-images", f)
+		if global.PathExists(cacheFile) {
+			return &LocalImageElement{File: cacheFile}, nil
+		}
+	}
+	if strings.HasSuffix(f, ".image") {
+		hash, err := hex.DecodeString(strings.TrimSuffix(f, ".image"))
+		if err == nil {
+			if b := cache.Image.Get(hash); b != nil {
+				return bot.readImageCache(b, sourceType)
 			}
-			return nil, errors.New("img size is 0")
 		}
-		if len(hash) != 16 {
-			return nil, errors.New("invalid hash")
+	}
+	exist := global.PathExists(rawPath)
+	if !exist && global.PathExists(path.Join(global.ImagePathOld, f)) {
+		exist = true
+		rawPath = path.Join(global.ImagePathOld, f)
+	}
+	if !exist {
+		if d["url"] != "" {
+			return bot.makeImageOrVideoElem(map[string]string{"file": d["url"]}, false, sourceType)
 		}
-		var rsp message.IMessageElement
-		if group {
-			rsp, err = bot.Client.QueryGroupImage(int64(rand.Uint32()), hash, size)
+		return nil, errors.New("invalid image")
+	}
+	if path.Ext(rawPath) != ".image" {
+		return &LocalImageElement{File: rawPath}, nil
+	}
+	b, err := os.ReadFile(rawPath)
+	if err != nil {
+		return nil, err
+	}
+	return bot.readImageCache(b, sourceType)
+}
+
+func (bot *CQBot) readImageCache(b []byte, sourceType MessageSourceType) (message.IMessageElement, error) {
+	var err error
+	if len(b) < 20 {
+		return nil, errors.New("invalid cache")
+	}
+	r := binary.NewReader(b)
+	hash := r.ReadBytes(16)
+	size := r.ReadInt32()
+	r.ReadString()
+	imageURL := r.ReadString()
+	if size == 0 {
+		if imageURL != "" {
+			return bot.makeImageOrVideoElem(map[string]string{"file": imageURL}, false, sourceType)
+		}
+		return nil, errors.New("img size is 0")
+	}
+	if len(hash) != 16 {
+		return nil, errors.New("invalid hash")
+	}
+	var rsp message.IMessageElement
+	if sourceType == MessageSourceGroup {
+		rsp, err = bot.Client.QueryGroupImage(int64(rand.Uint32()), hash, size)
+		goto ok
+	}
+	if sourceType == MessageSourceGuildChannel {
+		if len(bot.Client.GuildService.Guilds) == 0 {
+			err = errors.New("cannot query guild image: not any joined guild")
 			goto ok
 		}
-		rsp, err = bot.Client.QueryFriendImage(int64(rand.Uint32()), hash, size)
-	ok:
-		if err != nil {
-			if url != "" {
-				return bot.makeImageOrVideoElem(map[string]string{"file": url}, false, group)
-			}
-			return nil, err
-		}
-		return rsp, nil
+		guild := bot.Client.GuildService.Guilds[0]
+		rsp, err = bot.Client.GuildService.QueryImage(guild.GuildId, guild.Channels[0].ChannelId, hash, uint64(size))
+		goto ok
 	}
-	return nil, errors.New("invalid image")
+	rsp, err = bot.Client.QueryFriendImage(int64(rand.Uint32()), hash, size)
+ok:
+	if err != nil {
+		if imageURL != "" {
+			return bot.makeImageOrVideoElem(map[string]string{"file": imageURL}, false, sourceType)
+		}
+		return nil, err
+	}
+	return rsp, nil
+}
+
+func (bot *CQBot) readVideoCache(b []byte) message.IMessageElement {
+	r := binary.NewReader(b)
+	return &message.ShortVideoElement{ // todo 检查缓存是否有效
+		Md5:       r.ReadBytes(16),
+		ThumbMd5:  r.ReadBytes(16),
+		Size:      r.ReadInt32(),
+		ThumbSize: r.ReadInt32(),
+		Name:      r.ReadString(),
+		Uuid:      r.ReadAvailable(),
+	}
 }
 
 // makeShowPic 一种xml 方式发送的群消息图片
@@ -1240,39 +1426,26 @@ func (bot *CQBot) makeShowPic(elem message.IMessageElement, source string, brief
 	if brief == "" {
 		brief = "&#91;分享&#93;我看到一张很赞的图片，分享给你，快来看！"
 	}
-	if i, ok := elem.(*LocalImageElement); ok {
-		if !group {
-			gm, err := bot.UploadLocalImageAsPrivate(1, i)
-			if err != nil {
-				log.Warnf("警告: 好友消息 %v 消息图片上传失败: %v", 1, err)
-				return nil, err
-			}
-			suf = gm
-			xml = fmt.Sprintf(`<?xml version='1.0' encoding='UTF-8' standalone='yes' ?><msg serviceID="5" templateID="12345" action="" brief="%s" sourceMsgId="0" url="%s" flag="0" adverSign="0" multiMsgFlag="0"><item layout="0" advertiser_id="0" aid="0"><image uuid="%x" md5="%x" GroupFiledid="0" filesize="%d" local_path="%s" minWidth="%d" minHeight="%d" maxWidth="%d" maxHeight="%d" /></item><source name="%s" icon="%s" action="" appid="-1" /></msg>`, brief, "", gm.Md5, gm.Md5, len(i.Data), "", minWidth, minHeight, maxWidth, maxHeight, source, icon)
-		} else {
-			gm, err := bot.UploadLocalImageAsGroup(1, i)
-			if err != nil {
-				log.Warnf("警告: 群 %v 消息图片上传失败: %v", 1, err)
-				return nil, err
-			}
-			suf = gm
-			xml = fmt.Sprintf(`<?xml version='1.0' encoding='UTF-8' standalone='yes' ?><msg serviceID="5" templateID="12345" action="" brief="%s" sourceMsgId="0" url="%s" flag="0" adverSign="0" multiMsgFlag="0"><item layout="0" advertiser_id="0" aid="0"><image uuid="%x" md5="%x" GroupFiledid="0" filesize="%d" local_path="%s" minWidth="%d" minHeight="%d" maxWidth="%d" maxHeight="%d" /></item><source name="%s" icon="%s" action="" appid="-1" /></msg>`, brief, "", gm.Md5, gm.Md5, len(i.Data), "", minWidth, minHeight, maxWidth, maxHeight, source, icon)
+	if _, ok := elem.(*LocalImageElement); ok {
+		r := rand.Uint32()
+		e, err := bot.uploadMedia(elem, int64(r), group)
+		if err != nil {
+			log.Warnf("警告: 图片上传失败: %v", err)
+			return nil, err
 		}
+		elem = e
 	}
-
-	if i, ok := elem.(*message.GroupImageElement); ok {
+	switch i := elem.(type) {
+	case *message.GroupImageElement:
+		xml = fmt.Sprintf(`<?xml version='1.0' encoding='UTF-8' standalone='yes' ?><msg serviceID="5" templateID="12345" action="" brief="%s" sourceMsgId="0" url="%s" flag="0" adverSign="0" multiMsgFlag="0"><item layout="0" advertiser_id="0" aid="0"><image uuid="%x" md5="%x" GroupFiledid="0" filesize="%d" local_path="%s" minWidth="%d" minHeight="%d" maxWidth="%d" maxHeight="%d" /></item><source name="%s" icon="%s" action="" appid="-1" /></msg>`, brief, "", i.Md5, i.Md5, 0, "", minWidth, minHeight, maxWidth, maxHeight, source, icon)
+		suf = i
+	case *message.FriendImageElement:
 		xml = fmt.Sprintf(`<?xml version='1.0' encoding='UTF-8' standalone='yes' ?><msg serviceID="5" templateID="12345" action="" brief="%s" sourceMsgId="0" url="%s" flag="0" adverSign="0" multiMsgFlag="0"><item layout="0" advertiser_id="0" aid="0"><image uuid="%x" md5="%x" GroupFiledid="0" filesize="%d" local_path="%s" minWidth="%d" minHeight="%d" maxWidth="%d" maxHeight="%d" /></item><source name="%s" icon="%s" action="" appid="-1" /></msg>`, brief, "", i.Md5, i.Md5, 0, "", minWidth, minHeight, maxWidth, maxHeight, source, icon)
 		suf = i
 	}
-	if i, ok := elem.(*message.FriendImageElement); ok {
-		xml = fmt.Sprintf(`<?xml version='1.0' encoding='UTF-8' standalone='yes' ?><msg serviceID="5" templateID="12345" action="" brief="%s" sourceMsgId="0" url="%s" flag="0" adverSign="0" multiMsgFlag="0"><item layout="0" advertiser_id="0" aid="0"><image uuid="%x" md5="%x" GroupFiledid="0" filesize="%d" local_path="%s" minWidth="%d" minHeight="%d" maxWidth="%d" maxHeight="%d" /></item><source name="%s" icon="%s" action="" appid="-1" /></msg>`, brief, "", i.Md5, i.Md5, 0, "", minWidth, minHeight, maxWidth, maxHeight, source, icon)
-		suf = i
+	if xml == "" {
+		return nil, errors.New("生成xml图片消息失败")
 	}
-	if xml != "" {
-		// log.Warn(xml)
-		ret := []message.IMessageElement{suf}
-		ret = append(ret, message.NewRichXml(xml, 5))
-		return ret, nil
-	}
-	return nil, errors.New("生成xml图片消息失败")
+	ret := []message.IMessageElement{suf, message.NewRichXml(xml, 5)}
+	return ret, nil
 }
