@@ -2,7 +2,6 @@ package global
 
 import (
 	"bufio"
-	"bytes"
 	"compress/gzip"
 	"fmt"
 	"io"
@@ -13,22 +12,22 @@ import (
 	"strconv"
 	"strings"
 	"sync"
-	"time"
 
-	"github.com/guonaihong/gout"
 	"github.com/pkg/errors"
 
 	"github.com/tidwall/gjson"
+
+	"github.com/Mrs4s/go-cqhttp/internal/base"
 )
 
 var (
 	client = &http.Client{
 		Transport: &http.Transport{
 			Proxy: func(request *http.Request) (u *url.URL, e error) {
-				if Proxy == "" {
+				if base.Proxy == "" {
 					return http.ProxyFromEnvironment(request)
 				}
-				return url.Parse(Proxy)
+				return url.Parse(base.Proxy)
 			},
 			ForceAttemptHTTP2:   true,
 			MaxConnsPerHost:     0,
@@ -36,9 +35,6 @@ var (
 			MaxIdleConnsPerHost: 999,
 		},
 	}
-
-	// Proxy 存储Config.proxy_rewrite,用于设置代理
-	Proxy string
 
 	// ErrOverSize 响应主体过大时返回此错误
 	ErrOverSize = errors.New("oversize")
@@ -49,33 +45,19 @@ var (
 
 // GetBytes 对给定URL发送Get请求，返回响应主体
 func GetBytes(url string) ([]byte, error) {
-	req, err := http.NewRequest("GET", url, nil)
+	reader, err := HTTPGetReadCloser(url)
 	if err != nil {
 		return nil, err
 	}
-	req.Header["User-Agent"] = []string{UserAgent}
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-	if strings.Contains(resp.Header.Get("Content-Encoding"), "gzip") {
-		buffer := bytes.NewBuffer(body)
-		r, _ := gzip.NewReader(buffer)
-		defer r.Close()
-		unCom, err := ioutil.ReadAll(r)
-		return unCom, err
-	}
-	return body, nil
+	defer func() {
+		_ = reader.Close()
+	}()
+	return ioutil.ReadAll(reader)
 }
 
 // DownloadFile 将给定URL对应的文件下载至给定Path
 func DownloadFile(url, path string, limit int64, headers map[string]string) error {
-	file, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE, 0666)
+	file, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE, 0o666)
 	if err != nil {
 		return err
 	}
@@ -89,9 +71,6 @@ func DownloadFile(url, path string, limit int64, headers map[string]string) erro
 		req.Header.Set(k, v)
 	}
 
-	if _, ok := headers["User-Agent"]; !ok {
-		req.Header["User-Agent"] = []string{UserAgent}
-	}
 	resp, err := client.Do(req)
 	if err != nil {
 		return err
@@ -123,7 +102,7 @@ func DownloadFileMultiThreading(url, path string, limit int64, threadCount int, 
 	// 初始化分块或直接下载
 	initOrDownload := func() error {
 		copyStream := func(s io.ReadCloser) error {
-			file, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE, 0666)
+			file, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE, 0o666)
 			if err != nil {
 				return err
 			}
@@ -192,7 +171,7 @@ func DownloadFileMultiThreading(url, path string, limit int64, threadCount int, 
 	// 下载分块
 	downloadBlock := func(block *BlockMetaData) error {
 		req, _ := http.NewRequest("GET", url, nil)
-		file, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE, 0666)
+		file, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE, 0o666)
 		if err != nil {
 			return err
 		}
@@ -264,22 +243,6 @@ func DownloadFileMultiThreading(url, path string, limit int64, threadCount int, 
 	return lastErr
 }
 
-// GetSliderTicket 通过给定的验证链接raw和id,获取验证结果Ticket
-func GetSliderTicket(raw, id string) (string, error) {
-	var rsp string
-	if err := gout.POST("https://api.shkong.com/gocqhttpapi/task").SetJSON(gout.H{
-		"id":  id,
-		"url": raw,
-	}).SetTimeout(time.Second * 35).BindBody(&rsp).Do(); err != nil {
-		return "", err
-	}
-	g := gjson.Parse(rsp)
-	if g.Get("error").Str != "" {
-		return "", errors.New(g.Get("error").Str)
-	}
-	return g.Get("ticket").Str, nil
-}
-
 // QQMusicSongInfo 通过给定id在QQ音乐上查找曲目信息
 func QQMusicSongInfo(id string) (gjson.Result, error) {
 	d, err := GetBytes(`https://u.y.qq.com/cgi-bin/musicu.fcg?format=json&inCharset=utf8&outCharset=utf-8&notice=0&platform=yqq.json&needNewCode=0&data={%22comm%22:{%22ct%22:24,%22cv%22:0},%22songinfo%22:{%22method%22:%22get_song_detail_yqq%22,%22param%22:{%22song_type%22:0,%22song_mid%22:%22%22,%22song_id%22:` + id + `},%22module%22:%22music.pf_song_detail_svr%22}}`)
@@ -296,4 +259,49 @@ func NeteaseMusicSongInfo(id string) (gjson.Result, error) {
 		return gjson.Result{}, err
 	}
 	return gjson.ParseBytes(d).Get("songs.0"), nil
+}
+
+type gzipCloser struct {
+	f io.Closer
+	r *gzip.Reader
+}
+
+// NewGzipReadCloser 从 io.ReadCloser 创建 gunzip io.ReadCloser
+func NewGzipReadCloser(reader io.ReadCloser) (io.ReadCloser, error) {
+	gzipReader, err := gzip.NewReader(reader)
+	if err != nil {
+		return nil, err
+	}
+	return &gzipCloser{
+		f: reader,
+		r: gzipReader,
+	}, nil
+}
+
+// Read impls io.Reader
+func (g *gzipCloser) Read(p []byte) (n int, err error) {
+	return g.r.Read(p)
+}
+
+// Close impls io.Closer
+func (g *gzipCloser) Close() error {
+	_ = g.f.Close()
+	return g.r.Close()
+}
+
+// HTTPGetReadCloser 从 Http url 获取 io.ReadCloser
+func HTTPGetReadCloser(url string) (io.ReadCloser, error) {
+	req, err := http.NewRequest("GET", url, nil)
+	if err != nil {
+		return nil, err
+	}
+	req.Header["User-Agent"] = []string{UserAgent}
+	resp, err := client.Do(req)
+	if err != nil {
+		return nil, err
+	}
+	if strings.Contains(resp.Header.Get("Content-Encoding"), "gzip") {
+		return NewGzipReadCloser(resp.Body)
+	}
+	return resp.Body, err
 }
