@@ -2,6 +2,7 @@ package server
 
 import (
 	"bytes"
+	"encoding/json"
 	"fmt"
 	"net/http"
 	"runtime/debug"
@@ -10,42 +11,59 @@ import (
 	"sync"
 	"time"
 
-	"github.com/Mrs4s/go-cqhttp/coolq"
-	"github.com/Mrs4s/go-cqhttp/global"
-	"github.com/Mrs4s/go-cqhttp/global/config"
-
 	"github.com/Mrs4s/MiraiGo/utils"
-	"github.com/gorilla/websocket"
+	"github.com/RomiChan/websocket"
 	log "github.com/sirupsen/logrus"
 	"github.com/tidwall/gjson"
+	"gopkg.in/yaml.v3"
+
+	"github.com/Mrs4s/go-cqhttp/coolq"
+	"github.com/Mrs4s/go-cqhttp/global"
+	"github.com/Mrs4s/go-cqhttp/modules/api"
+	"github.com/Mrs4s/go-cqhttp/modules/config"
+	"github.com/Mrs4s/go-cqhttp/modules/filter"
 )
 
 type webSocketServer struct {
 	bot  *coolq.CQBot
-	conf *config.WebsocketServer
+	conf *WebsocketServer
 
-	eventConn      []*webSocketConn
-	eventConnMutex sync.Mutex
-	token          string
-	handshake      string
-	filter         string
+	mu        sync.Mutex
+	eventConn []*wsConn
+
+	token     string
+	handshake string
+	filter    string
 }
 
-// WebSocketClient WebSocket客户端实例
-type WebSocketClient struct {
-	bot  *coolq.CQBot
-	conf *config.WebsocketReverse
+// websocketClient WebSocket客户端实例
+type websocketClient struct {
+	bot       *coolq.CQBot
+	mu        sync.Mutex
+	universal *wsConn
+	event     *wsConn
 
-	universalConn *webSocketConn
-	eventConn     *webSocketConn
-	token         string
-	filter        string
+	token             string
+	filter            string
+	reconnectInterval time.Duration
+	limiter           api.Handler
 }
 
-type webSocketConn struct {
-	*websocket.Conn
-	sync.Mutex
-	apiCaller *apiCaller
+type wsConn struct {
+	mu        sync.Mutex
+	conn      *websocket.Conn
+	apiCaller *api.Caller
+}
+
+func (c *wsConn) WriteText(b []byte) error {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	_ = c.conn.SetWriteDeadline(time.Now().Add(time.Second * 15))
+	return c.conn.WriteMessage(websocket.TextMessage, b)
+}
+
+func (c *wsConn) Close() error {
+	return c.conn.Close()
 }
 
 var upgrader = websocket.Upgrader{
@@ -54,158 +72,194 @@ var upgrader = websocket.Upgrader{
 	},
 }
 
-// RunWebSocketServer 运行一个正向WS server
-func RunWebSocketServer(b *coolq.CQBot, conf *config.WebsocketServer) {
-	if conf.Disabled {
+const wsDefault = `  # 正向WS设置
+  - ws:
+      # 正向WS服务器监听地址
+      host: 127.0.0.1
+      # 正向WS服务器监听端口
+      port: 6700
+      middlewares:
+        <<: *default # 引用默认中间件
+`
+
+const wsReverseDefault = `  # 反向WS设置
+  - ws-reverse:
+      # 反向WS Universal 地址
+      # 注意 设置了此项地址后下面两项将会被忽略
+      universal: ws://your_websocket_universal.server
+      # 反向WS API 地址
+      api: ws://your_websocket_api.server
+      # 反向WS Event 地址
+      event: ws://your_websocket_event.server
+      # 重连间隔 单位毫秒
+      reconnect-interval: 3000
+      middlewares:
+        <<: *default # 引用默认中间件
+`
+
+// WebsocketServer 正向WS相关配置
+type WebsocketServer struct {
+	Disabled bool   `yaml:"disabled"`
+	Host     string `yaml:"host"`
+	Port     int    `yaml:"port"`
+
+	MiddleWares `yaml:"middlewares"`
+}
+
+// WebsocketReverse 反向WS相关配置
+type WebsocketReverse struct {
+	Disabled          bool   `yaml:"disabled"`
+	Universal         string `yaml:"universal"`
+	API               string `yaml:"api"`
+	Event             string `yaml:"event"`
+	ReconnectInterval int    `yaml:"reconnect-interval"`
+
+	MiddleWares `yaml:"middlewares"`
+}
+
+func init() {
+	config.AddServer(&config.Server{
+		Brief:   "正向 Websocket 通信",
+		Default: wsDefault,
+	})
+	config.AddServer(&config.Server{
+		Brief:   "反向 Websocket 通信",
+		Default: wsReverseDefault,
+	})
+}
+
+// runWSServer 运行一个正向WS server
+func runWSServer(b *coolq.CQBot, node yaml.Node) {
+	var conf WebsocketServer
+	switch err := node.Decode(&conf); {
+	case err != nil:
+		log.Warn("读取正向Websocket配置失败 :", err)
+		fallthrough
+	case conf.Disabled:
 		return
 	}
-	s := new(webSocketServer)
-	s.conf = conf
-	s.bot = b
-	s.token = conf.AccessToken
-	s.filter = conf.Filter
-	addFilter(s.filter)
+
+	s := &webSocketServer{
+		bot:    b,
+		conf:   &conf,
+		token:  conf.AccessToken,
+		filter: conf.Filter,
+	}
+	filter.Add(s.filter)
 	addr := fmt.Sprintf("%s:%d", conf.Host, conf.Port)
 	s.handshake = fmt.Sprintf(`{"_post_method":2,"meta_event_type":"lifecycle","post_type":"meta_event","self_id":%d,"sub_type":"connect","time":%d}`,
 		b.Client.Uin, time.Now().Unix())
 	b.OnEventPush(s.onBotPushEvent)
-	http.HandleFunc("/event", s.event)
-	http.HandleFunc("/api", s.api)
-	http.HandleFunc("/", s.any)
-	go func() {
-		log.Infof("CQ WebSocket 服务器已启动: %v", addr)
-		log.Fatal(http.ListenAndServe(addr, nil))
-	}()
+	mux := http.ServeMux{}
+	mux.HandleFunc("/event", s.event)
+	mux.HandleFunc("/api", s.api)
+	mux.HandleFunc("/", s.any)
+	log.Infof("CQ WebSocket 服务器已启动: %v", addr)
+	log.Fatal(http.ListenAndServe(addr, &mux))
 }
 
-// RunWebSocketClient 运行一个正向WS client
-func RunWebSocketClient(b *coolq.CQBot, conf *config.WebsocketReverse) {
-	if conf.Disabled {
+// runWSClient 运行一个反向向WS client
+func runWSClient(b *coolq.CQBot, node yaml.Node) {
+	var conf WebsocketReverse
+	switch err := node.Decode(&conf); {
+	case err != nil:
+		log.Warn("读取反向Websocket配置失败 :", err)
+		fallthrough
+	case conf.Disabled:
 		return
 	}
-	c := new(WebSocketClient)
-	c.bot = b
-	c.conf = conf
-	c.token = conf.AccessToken
-	c.filter = conf.Filter
-	addFilter(c.filter)
-	if c.conf.Universal != "" {
-		c.connectUniversal()
+
+	c := &websocketClient{
+		bot:    b,
+		token:  conf.AccessToken,
+		filter: conf.Filter,
+	}
+	filter.Add(c.filter)
+	if conf.ReconnectInterval != 0 {
+		c.reconnectInterval = time.Duration(conf.ReconnectInterval) * time.Millisecond
+	}
+	if conf.RateLimit.Enabled {
+		c.limiter = rateLimit(conf.RateLimit.Frequency, conf.RateLimit.Bucket)
+	}
+
+	if conf.Universal != "" {
+		c.connect("Universal", conf.Universal, &c.universal)
+		c.bot.OnEventPush(c.onBotPushEvent("Universal", conf.Universal, &c.universal))
+		return // 连接到 Universal 后， 不再连接其他
+	}
+	if conf.API != "" {
+		c.connect("API", conf.API, nil)
+	}
+	if conf.Event != "" {
+		c.connect("Event", conf.Event, &c.event)
+		c.bot.OnEventPush(c.onBotPushEvent("Event", conf.Event, &c.event))
+	}
+}
+
+func (c *websocketClient) connect(typ, url string, conptr **wsConn) {
+	log.Infof("开始尝试连接到反向WebSocket %s服务器: %v", typ, url)
+	header := http.Header{
+		"X-Client-Role": []string{typ},
+		"X-Self-ID":     []string{strconv.FormatInt(c.bot.Client.Uin, 10)},
+		"User-Agent":    []string{"CQHttp/4.15.0"},
+	}
+	if c.token != "" {
+		header["Authorization"] = []string{"Token " + c.token}
+	}
+	conn, _, err := websocket.DefaultDialer.Dial(url, header) // nolint
+	if err != nil {
+		log.Warnf("连接到反向WebSocket %s服务器 %v 时出现错误: %v", typ, url, err)
+		if c.reconnectInterval != 0 {
+			time.Sleep(c.reconnectInterval)
+			c.connect(typ, url, conptr)
+		}
+		return
+	}
+
+	switch typ {
+	case "Event", "Universal":
+		handshake := fmt.Sprintf(`{"meta_event_type":"lifecycle","post_type":"meta_event","self_id":%d,"sub_type":"connect","time":%d}`, c.bot.Client.Uin, time.Now().Unix())
+		err = conn.WriteMessage(websocket.TextMessage, []byte(handshake))
+		if err != nil {
+			log.Warnf("反向WebSocket 握手时出现错误: %v", err)
+		}
+	}
+
+	log.Infof("已连接到反向WebSocket %s服务器 %v", typ, url)
+
+	var wrappedConn *wsConn
+	if conptr != nil && *conptr != nil {
+		wrappedConn = *conptr
 	} else {
-		if c.conf.API != "" {
-			c.connectAPI()
-		}
-		if c.conf.Event != "" {
-			c.connectEvent()
+		wrappedConn = new(wsConn)
+		if conptr != nil {
+			*conptr = wrappedConn
 		}
 	}
-	c.bot.OnEventPush(c.onBotPushEvent)
+
+	wrappedConn.conn = conn
+	wrappedConn.apiCaller = api.NewCaller(c.bot)
+	if c.limiter != nil {
+		wrappedConn.apiCaller.Use(c.limiter)
+	}
+
+	if typ != "Event" {
+		go c.listenAPI(typ, url, wrappedConn)
+	}
 }
 
-func (c *WebSocketClient) connectAPI() {
-	log.Infof("开始尝试连接到反向WebSocket API服务器: %v", c.conf.API)
-	header := http.Header{
-		"X-Client-Role": []string{"API"},
-		"X-Self-ID":     []string{strconv.FormatInt(c.bot.Client.Uin, 10)},
-		"User-Agent":    []string{"CQHttp/4.15.0"},
-	}
-	if c.token != "" {
-		header["Authorization"] = []string{"Token " + c.token}
-	}
-	conn, _, err := websocket.DefaultDialer.Dial(c.conf.API, header) // nolint
-	if err != nil {
-		log.Warnf("连接到反向WebSocket API服务器 %v 时出现错误: %v", c.conf.API, err)
-		if c.conf.ReconnectInterval != 0 {
-			time.Sleep(time.Millisecond * time.Duration(c.conf.ReconnectInterval))
-			c.connectAPI()
-		}
-		return
-	}
-	log.Infof("已连接到反向WebSocket API服务器 %v", c.conf.API)
-	wrappedConn := &webSocketConn{Conn: conn, apiCaller: newAPICaller(c.bot)}
-	if c.conf.RateLimit.Enabled {
-		wrappedConn.apiCaller.use(rateLimit(c.conf.RateLimit.Frequency, c.conf.RateLimit.Bucket))
-	}
-	go c.listenAPI(wrappedConn, false)
-}
-
-func (c *WebSocketClient) connectEvent() {
-	log.Infof("开始尝试连接到反向WebSocket Event服务器: %v", c.conf.Event)
-	header := http.Header{
-		"X-Client-Role": []string{"Event"},
-		"X-Self-ID":     []string{strconv.FormatInt(c.bot.Client.Uin, 10)},
-		"User-Agent":    []string{"CQHttp/4.15.0"},
-	}
-	if c.token != "" {
-		header["Authorization"] = []string{"Token " + c.token}
-	}
-	conn, _, err := websocket.DefaultDialer.Dial(c.conf.Event, header) // nolint
-	if err != nil {
-		log.Warnf("连接到反向WebSocket Event服务器 %v 时出现错误: %v", c.conf.Event, err)
-		if c.conf.ReconnectInterval != 0 {
-			time.Sleep(time.Millisecond * time.Duration(c.conf.ReconnectInterval))
-			c.connectEvent()
-		}
-		return
-	}
-
-	handshake := fmt.Sprintf(`{"meta_event_type":"lifecycle","post_type":"meta_event","self_id":%d,"sub_type":"connect","time":%d}`,
-		c.bot.Client.Uin, time.Now().Unix())
-	err = conn.WriteMessage(websocket.TextMessage, []byte(handshake))
-	if err != nil {
-		log.Warnf("反向WebSocket 握手时出现错误: %v", err)
-	}
-
-	log.Infof("已连接到反向WebSocket Event服务器 %v", c.conf.Event)
-	c.eventConn = &webSocketConn{Conn: conn, apiCaller: newAPICaller(c.bot)}
-}
-
-func (c *WebSocketClient) connectUniversal() {
-	log.Infof("开始尝试连接到反向WebSocket Universal服务器: %v", c.conf.Universal)
-	header := http.Header{
-		"X-Client-Role": []string{"Universal"},
-		"X-Self-ID":     []string{strconv.FormatInt(c.bot.Client.Uin, 10)},
-		"User-Agent":    []string{"CQHttp/4.15.0"},
-	}
-	if c.token != "" {
-		header["Authorization"] = []string{"Token " + c.token}
-	}
-	conn, _, err := websocket.DefaultDialer.Dial(c.conf.Universal, header) // nolint
-	if err != nil {
-		log.Warnf("连接到反向WebSocket Universal服务器 %v 时出现错误: %v", c.conf.Universal, err)
-		if c.conf.ReconnectInterval != 0 {
-			time.Sleep(time.Millisecond * time.Duration(c.conf.ReconnectInterval))
-			c.connectUniversal()
-		}
-		return
-	}
-	handshake := fmt.Sprintf(`{"meta_event_type":"lifecycle","post_type":"meta_event","self_id":%d,"sub_type":"connect","time":%d}`,
-		c.bot.Client.Uin, time.Now().Unix())
-	err = conn.WriteMessage(websocket.TextMessage, []byte(handshake))
-	if err != nil {
-		log.Warnf("反向WebSocket 握手时出现错误: %v", err)
-	}
-
-	wrappedConn := &webSocketConn{Conn: conn, apiCaller: newAPICaller(c.bot)}
-	if c.conf.RateLimit.Enabled {
-		wrappedConn.apiCaller.use(rateLimit(c.conf.RateLimit.Frequency, c.conf.RateLimit.Bucket))
-	}
-	go c.listenAPI(wrappedConn, true)
-	c.universalConn = wrappedConn
-}
-
-func (c *WebSocketClient) listenAPI(conn *webSocketConn, u bool) {
+func (c *websocketClient) listenAPI(typ, url string, conn *wsConn) {
 	defer func() { _ = conn.Close() }()
 	for {
 		buffer := global.NewBuffer()
-		t, reader, err := conn.NextReader()
+		t, reader, err := conn.conn.NextReader()
 		if err != nil {
-			log.Warnf("监听反向WS API时出现错误: %v", err)
+			log.Warnf("监听反向WS %s时出现错误: %v", typ, err)
 			break
 		}
 		_, err = buffer.ReadFrom(reader)
 		if err != nil {
-			log.Warnf("监听反向WS API时出现错误: %v", err)
+			log.Warnf("监听反向WS %s时出现错误: %v", typ, err)
 			break
 		}
 		if t == websocket.TextMessage {
@@ -217,82 +271,51 @@ func (c *WebSocketClient) listenAPI(conn *webSocketConn, u bool) {
 			global.PutBuffer(buffer)
 		}
 	}
-	if c.conf.ReconnectInterval != 0 {
-		time.Sleep(time.Millisecond * time.Duration(c.conf.ReconnectInterval))
-		if !u {
-			go c.connectAPI()
+	if c.reconnectInterval != 0 {
+		time.Sleep(c.reconnectInterval)
+		if typ == "API" { // Universal 不重连，避免多次重连
+			go c.connect(typ, url, nil)
 		}
 	}
 }
 
-func (c *WebSocketClient) onBotPushEvent(m *bytes.Buffer) {
-	filter := findFilter(c.filter)
-	if filter != nil && !filter.Eval(gjson.Parse(utils.B2S(m.Bytes()))) {
-		log.Debugf("上报Event %v 到 WS客户端 时被过滤.", utils.B2S(m.Bytes()))
-		return
-	}
+func (c *websocketClient) onBotPushEvent(typ, url string, conn **wsConn) func(e *coolq.Event) {
+	return func(e *coolq.Event) {
+		c.mu.Lock()
+		defer c.mu.Unlock()
 
-	if c.eventConn != nil {
-		log.Debugf("向WS服务器 %v 推送Event: %v", c.eventConn.RemoteAddr().String(), utils.B2S(m.Bytes()))
-		conn := c.eventConn
-		conn.Lock()
-		defer conn.Unlock()
-		_ = c.eventConn.SetWriteDeadline(time.Now().Add(time.Second * 15))
-		if err := c.eventConn.WriteMessage(websocket.TextMessage, m.Bytes()); err != nil {
-			log.Warnf("向WS服务器 %v 推送Event时出现错误: %v", c.eventConn.RemoteAddr().String(), err)
-			_ = c.eventConn.Close()
-			if c.conf.ReconnectInterval != 0 {
-				time.Sleep(time.Millisecond * time.Duration(c.conf.ReconnectInterval))
-				c.connectEvent()
-			}
+		flt := filter.Find(c.filter)
+		if flt != nil && !flt.Eval(gjson.Parse(e.JSONString())) {
+			log.Debugf("上报Event %s 到 WS服务器 时被过滤.", e.JSONBytes())
+			return
 		}
-	}
-	if c.universalConn != nil {
-		log.Debugf("向WS服务器 %v 推送Event: %v", c.universalConn.RemoteAddr().String(), utils.B2S(m.Bytes()))
-		conn := c.universalConn
-		conn.Lock()
-		defer conn.Unlock()
-		_ = c.universalConn.SetWriteDeadline(time.Now().Add(time.Second * 15))
-		if err := c.universalConn.WriteMessage(websocket.TextMessage, m.Bytes()); err != nil {
-			log.Warnf("向WS服务器 %v 推送Event时出现错误: %v", c.universalConn.RemoteAddr().String(), err)
-			_ = c.universalConn.Close()
-			if c.conf.ReconnectInterval != 0 {
-				time.Sleep(time.Millisecond * time.Duration(c.conf.ReconnectInterval))
-				c.connectUniversal()
-			}
-		}
-	}
-}
 
-func (s *webSocketServer) auth(r *http.Request) (bool, int) {
-	if s.token != "" { // s.token == s.conf.AccessToken
-		var auth string
-		if auth = r.URL.Query().Get("access_token"); auth == "" {
-			headAuth := strings.SplitN(r.Header.Get("Authorization"), " ", 2)
-			if len(headAuth) != 2 || headAuth[1] == "" {
-				return false, 401
+		log.Debugf("向反向WS %s服务器推送Event: %s", typ, e.JSONBytes())
+		if err := (*conn).WriteText(e.JSONBytes()); err != nil {
+			log.Warnf("向反向WS %s服务器推送 Event 时出现错误: %v", typ, err)
+			_ = (*conn).Close()
+			if c.reconnectInterval != 0 {
+				time.Sleep(c.reconnectInterval)
+				c.connect(typ, url, conn)
 			}
-			auth = headAuth[1]
-		}
-		if auth != s.token {
-			log.Warnf("已拒绝 %v 的 WebSocket 请求: Token鉴权失败", r.RemoteAddr)
-			return false, 403
 		}
 	}
-	return true, 0
 }
 
 func (s *webSocketServer) event(w http.ResponseWriter, r *http.Request) {
-	isAuth, errReason := s.auth(r)
-	if !isAuth {
-		w.WriteHeader(errReason)
+	status := checkAuth(r, s.token)
+	if status != http.StatusOK {
+		log.Warnf("已拒绝 %v 的 WebSocket 请求: Token鉴权失败(code:%d)", r.RemoteAddr, status)
+		w.WriteHeader(status)
 		return
 	}
+
 	c, err := upgrader.Upgrade(w, r, nil)
 	if err != nil {
 		log.Warnf("处理 WebSocket 请求时出现错误: %v", err)
 		return
 	}
+
 	err = c.WriteMessage(websocket.TextMessage, []byte(s.handshake))
 	if err != nil {
 		log.Warnf("WebSocket 握手时出现错误: %v", err)
@@ -301,66 +324,71 @@ func (s *webSocketServer) event(w http.ResponseWriter, r *http.Request) {
 	}
 
 	log.Infof("接受 WebSocket 连接: %v (/event)", r.RemoteAddr)
-
-	conn := &webSocketConn{Conn: c, apiCaller: newAPICaller(s.bot)}
-
-	s.eventConnMutex.Lock()
+	conn := &wsConn{conn: c, apiCaller: api.NewCaller(s.bot)}
+	s.mu.Lock()
 	s.eventConn = append(s.eventConn, conn)
-	s.eventConnMutex.Unlock()
+	s.mu.Unlock()
 }
 
 func (s *webSocketServer) api(w http.ResponseWriter, r *http.Request) {
-	isAuth, errReason := s.auth(r)
-	if !isAuth {
-		w.WriteHeader(errReason)
+	status := checkAuth(r, s.token)
+	if status != http.StatusOK {
+		log.Warnf("已拒绝 %v 的 WebSocket 请求: Token鉴权失败(code:%d)", r.RemoteAddr, status)
+		w.WriteHeader(status)
 		return
 	}
+
 	c, err := upgrader.Upgrade(w, r, nil)
 	if err != nil {
 		log.Warnf("处理 WebSocket 请求时出现错误: %v", err)
 		return
 	}
+
 	log.Infof("接受 WebSocket 连接: %v (/api)", r.RemoteAddr)
-	conn := &webSocketConn{Conn: c, apiCaller: newAPICaller(s.bot)}
+	conn := &wsConn{conn: c, apiCaller: api.NewCaller(s.bot)}
 	if s.conf.RateLimit.Enabled {
-		conn.apiCaller.use(rateLimit(s.conf.RateLimit.Frequency, s.conf.RateLimit.Bucket))
+		conn.apiCaller.Use(rateLimit(s.conf.RateLimit.Frequency, s.conf.RateLimit.Bucket))
 	}
-	go s.listenAPI(conn)
+	s.listenAPI(conn)
 }
 
 func (s *webSocketServer) any(w http.ResponseWriter, r *http.Request) {
-	isAuth, errReason := s.auth(r)
-	if !isAuth {
-		w.WriteHeader(errReason)
+	status := checkAuth(r, s.token)
+	if status != http.StatusOK {
+		log.Warnf("已拒绝 %v 的 WebSocket 请求: Token鉴权失败(code:%d)", r.RemoteAddr, status)
+		w.WriteHeader(status)
 		return
 	}
+
 	c, err := upgrader.Upgrade(w, r, nil)
 	if err != nil {
 		log.Warnf("处理 WebSocket 请求时出现错误: %v", err)
 		return
 	}
+
 	err = c.WriteMessage(websocket.TextMessage, []byte(s.handshake))
 	if err != nil {
 		log.Warnf("WebSocket 握手时出现错误: %v", err)
 		_ = c.Close()
 		return
 	}
+
 	log.Infof("接受 WebSocket 连接: %v (/)", r.RemoteAddr)
-	conn := &webSocketConn{Conn: c, apiCaller: newAPICaller(s.bot)}
+	conn := &wsConn{conn: c, apiCaller: api.NewCaller(s.bot)}
 	if s.conf.RateLimit.Enabled {
-		conn.apiCaller.use(rateLimit(s.conf.RateLimit.Frequency, s.conf.RateLimit.Bucket))
+		conn.apiCaller.Use(rateLimit(s.conf.RateLimit.Frequency, s.conf.RateLimit.Bucket))
 	}
-	s.eventConnMutex.Lock()
+	s.mu.Lock()
 	s.eventConn = append(s.eventConn, conn)
-	s.eventConnMutex.Unlock()
+	s.mu.Unlock()
 	s.listenAPI(conn)
 }
 
-func (s *webSocketServer) listenAPI(c *webSocketConn) {
+func (s *webSocketServer) listenAPI(c *wsConn) {
 	defer func() { _ = c.Close() }()
 	for {
 		buffer := global.NewBuffer()
-		t, reader, err := c.NextReader()
+		t, reader, err := c.conn.NextReader()
 		if err != nil {
 			break
 		}
@@ -380,52 +408,54 @@ func (s *webSocketServer) listenAPI(c *webSocketConn) {
 	}
 }
 
-func (c *webSocketConn) handleRequest(_ *coolq.CQBot, payload []byte) {
+func (c *wsConn) handleRequest(_ *coolq.CQBot, payload []byte) {
 	defer func() {
 		if err := recover(); err != nil {
 			log.Printf("处置WS命令时发生无法恢复的异常：%v\n%s", err, debug.Stack())
 			_ = c.Close()
 		}
 	}()
-	j := gjson.ParseBytes(payload)
-	t := strings.ReplaceAll(j.Get("action").Str, "_async", "")
+	j := gjson.Parse(utils.B2S(payload))
+	t := strings.TrimSuffix(j.Get("action").Str, "_async")
 	log.Debugf("WS接收到API调用: %v 参数: %v", t, j.Get("params").Raw)
-	ret := c.apiCaller.callAPI(t, j.Get("params"))
+	ret := c.apiCaller.Call(t, j.Get("params"))
 	if j.Get("echo").Exists() {
 		ret["echo"] = j.Get("echo").Value()
 	}
-	c.Lock()
-	defer c.Unlock()
-	_ = c.WriteJSON(ret)
+
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	_ = c.conn.SetWriteDeadline(time.Now().Add(time.Second * 15))
+	writer, _ := c.conn.NextWriter(websocket.TextMessage)
+	_ = json.NewEncoder(writer).Encode(ret)
+	_ = writer.Close()
 }
 
-func (s *webSocketServer) onBotPushEvent(m *bytes.Buffer) {
-	s.eventConnMutex.Lock()
-	defer s.eventConnMutex.Unlock()
-
-	filter := findFilter(s.filter)
-	if filter != nil && !filter.Eval(gjson.Parse(utils.B2S(m.Bytes()))) {
-		log.Debugf("上报Event %v 到 WS客户端 时被过滤.", utils.B2S(m.Bytes()))
+func (s *webSocketServer) onBotPushEvent(e *coolq.Event) {
+	flt := filter.Find(s.filter)
+	if flt != nil && !flt.Eval(gjson.Parse(e.JSONString())) {
+		log.Debugf("上报Event %s 到 WS客户端 时被过滤.", e.JSONBytes())
 		return
 	}
 
-	for i, l := 0, len(s.eventConn); i < l; i++ {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	j := 0
+	for i := 0; i < len(s.eventConn); i++ {
 		conn := s.eventConn[i]
-		log.Debugf("向WS客户端 %v 推送Event: %v", conn.RemoteAddr().String(), utils.B2S(m.Bytes()))
-		conn.Lock()
-		if err := conn.WriteMessage(websocket.TextMessage, m.Bytes()); err != nil {
+		log.Debugf("向WS客户端推送Event: %s", e.JSONBytes())
+		if err := conn.WriteText(e.JSONBytes()); err != nil {
 			_ = conn.Close()
-			next := i + 1
-			if next >= l {
-				next = l - 1
-			}
-			s.eventConn[i], s.eventConn[next] = s.eventConn[next], s.eventConn[i]
-			s.eventConn = append(s.eventConn[:next], s.eventConn[next+1:]...)
-			i--
-			l--
 			conn = nil
 			continue
 		}
-		conn.Unlock()
+		if i != j {
+			// i != j means that some connection has been closed.
+			// use an in-place removal to avoid copying.
+			s.eventConn[j] = conn
+		}
+		j++
 	}
+	s.eventConn = s.eventConn[:j]
 }
